Crate hotshot_types

Types and Traits for the HotShot consensus module

Modules

consensus

Provides the core consensus types

data

Provides types useful for representing HotShot’s data structures

error

Error type for HotShot

event

Events that a HotShot instance can emit

light_client

Types and structs associated with light client state

message

Network message types

qc

Implementation for BitVectorQC that uses BLS signature + Bit vector. See more details in hotshot paper.

signature_key

Types and structs for the hotshot signature keys

simple_certificate

Implementations of the simple certificate type. Used for Quorum, DA, and Timeout Certificates

simple_vote

Implementations of the simple vote types.

stake_table

Types and structs related to the stake table

traits

Common traits for the HotShot protocol

utils

Utility functions, type aliases, helper structs and enum definitions.

vote

Vote, Accumulator, and Certificate Types

Structs

HotShotConfig

Holds configuration for a HotShot

ValidatorConfig

config for validator, including public key, private key, stake value

Enums

ExecutionType

the type of consensus to run. Either: wait for a signal to start a view, or constantly run you almost always want continuous incremental is just for testing

Constants

_DERIVE_Display_FOR_ValidatorConfig 🔒

_DERIVE_core_fmt_Debug_FOR_HotShotConfig 🔒

Functions

assert_future

yoinked from futures crate

boxed_sync

yoinked from futures crate, adds sync bound that we need

Type Aliases

BoxSyncFuture

Pinned future that is Send and Sync