Module hotshot_types::consensus

Provides the core consensus types

Re-exports

pub use crate::traits::node_implementation::ViewQueue;

pub use crate::utils::View;

pub use crate::utils::ViewInner;

Structs

Consensus

A reference to the consensus algorithm

ConsensusMetrics

The wrapper with a string name for the networking metrics

ConsensusMetricsValue

Contains several ConsensusMetrics that we’re interested in from the consensus interfaces

InnerConsensusMetrics

the set of counters and gauges for the networking metrics

Constants

_DERIVE_Display_FOR_InnerConsensusMetrics 🔒

_DERIVE_core_fmt_Debug_FOR_Consensus 🔒

Type Aliases

CommitmentMap 🔒

A type alias for HashMap<Commitment<T>, T>