Struct hotshot_types::consensus::Consensus

pub struct Consensus<TYPES: NodeType> {

    validated_state_map: BTreeMap<TYPES::View, View<TYPES>>,
    vid_shares: VidShares<TYPES>,
    saved_da_certs: HashMap<TYPES::View, DaCertificate2<TYPES>>,
    cur_view: TYPES::View,
    cur_epoch: Option<TYPES::Epoch>,
    last_proposals: BTreeMap<TYPES::View, Proposal<TYPES, QuorumProposalWrapper<TYPES>>>,
    last_decided_view: TYPES::View,
    locked_view: TYPES::View,
    saved_leaves: CommitmentMap<Leaf2<TYPES>>,
    last_actions: HotShotActionViews<TYPES::View>,
    saved_payloads: BTreeMap<TYPES::View, Arc<TYPES::BlockPayload>>,
    high_qc: QuorumCertificate2<TYPES>,
    next_epoch_high_qc: Option<NextEpochQuorumCertificate2<TYPES>>,
    pub metrics: Arc<ConsensusMetricsValue>,
    pub epoch_height: u64,
    pub drb_seeds_and_results: DrbSeedsAndResults<TYPES>,

}

A reference to the consensus algorithm

This will contain the state of all rounds.

Fields
validated_state_map: BTreeMap<TYPES::View, View<TYPES>>

The validated states that are currently loaded in memory.

vid_shares: VidShares<TYPES>

All the VID shares we’ve received for current and future views.

saved_da_certs: HashMap<TYPES::View, DaCertificate2<TYPES>>

All the DA certs we’ve received for current and future views. view -> DA cert

cur_view: TYPES::View

View number that is currently on.

cur_epoch: Option<TYPES::Epoch>

Epoch number that is currently on.

last_proposals: BTreeMap<TYPES::View, Proposal<TYPES, QuorumProposalWrapper<TYPES>>>

Last proposals we sent out, None if we haven’t proposed yet. Prevents duplicate proposals, and can be served to those trying to catchup

last_decided_view: TYPES::View

last view had a successful decide event

locked_view: TYPES::View

The locked_qc view number

saved_leaves: CommitmentMap<Leaf2<TYPES>>

Map of leaf hash -> leaf

contains undecided leaves

includes the MOST RECENT decided leaf

last_actions: HotShotActionViews<TYPES::View>

Bundle of views which we performed the most recent action visibible to the network. Actions are votes and proposals for DA and Quorum

saved_payloads: BTreeMap<TYPES::View, Arc<TYPES::BlockPayload>>

Saved payloads.

Encoded transactions for every view if we got a payload for that view.

high_qc: QuorumCertificate2<TYPES>

the highqc per spec

next_epoch_high_qc: Option<NextEpochQuorumCertificate2<TYPES>>

The high QC for the next epoch

metrics: Arc<ConsensusMetricsValue>

A reference to the metrics trait

epoch_height: u64

Number of blocks in an epoch, zero means there are no epochs

drb_seeds_and_results: DrbSeedsAndResults<TYPES>

Tables for the DRB seeds and results.

Implementations

impl<TYPES: NodeType> Consensus<TYPES>

pub fn new( validated_state_map: BTreeMap<TYPES::View, View<TYPES>>, vid_shares: Option<VidShares<TYPES>>, cur_view: TYPES::View, cur_epoch: Option<TYPES::Epoch>, locked_view: TYPES::View, last_decided_view: TYPES::View, last_actioned_view: TYPES::View, last_proposals: BTreeMap<TYPES::View, Proposal<TYPES, QuorumProposalWrapper<TYPES>>>, saved_leaves: CommitmentMap<Leaf2<TYPES>>, saved_payloads: BTreeMap<TYPES::View, Arc<TYPES::BlockPayload>>, high_qc: QuorumCertificate2<TYPES>, next_epoch_high_qc: Option<NextEpochQuorumCertificate2<TYPES>>, metrics: Arc<ConsensusMetricsValue>, epoch_height: u64, ) -> Self

Constructor.

pub fn cur_view(&self) -> TYPES::View

Get the current view.

pub fn cur_epoch(&self) -> Option<TYPES::Epoch>

Get the current epoch.

pub fn last_decided_view(&self) -> TYPES::View

Get the last decided view.

pub fn locked_view(&self) -> TYPES::View

Get the locked view.

pub fn high_qc(&self) -> &QuorumCertificate2<TYPES>

Get the high QC.

pub fn next_epoch_high_qc(&self) -> Option<&NextEpochQuorumCertificate2<TYPES>>

Get the next epoch high QC.

pub fn validated_state_map(&self) -> &BTreeMap<TYPES::View, View<TYPES>>

Get the validated state map.

pub fn saved_leaves(&self) -> &CommitmentMap<Leaf2<TYPES>>

Get the saved leaves.

pub fn saved_payloads(&self) -> &BTreeMap<TYPES::View, Arc<TYPES::BlockPayload>>

Get the saved payloads.

pub fn vid_shares(&self) -> &VidShares<TYPES>

Get the vid shares.

pub fn saved_da_certs(&self) -> &HashMap<TYPES::View, DaCertificate2<TYPES>>

Get the saved DA certs.

pub fn last_proposals( &self, ) -> &BTreeMap<TYPES::View, Proposal<TYPES, QuorumProposalWrapper<TYPES>>>

Get the map of our recent proposals

pub fn update_view(&mut self, view_number: TYPES::View) -> Result<()>

Update the current view.

Errors

Can return an error when the new view_number is not higher than the existing view number.

pub fn parent_leaf_info( &self, leaf: &Leaf2<TYPES>, public_key: &TYPES::SignatureKey, ) -> Option<LeafInfo<TYPES>>

Get the parent Leaf Info from a given leaf and our public key. Returns None if we don’t have the data in out state

pub fn update_epoch(&mut self, epoch_number: TYPES::Epoch) -> Result<()>

Update the current epoch.

Errors

Can return an error when the new epoch_number is not higher than the existing epoch number.

pub fn update_action( &mut self, action: HotShotAction, view: TYPES::View, ) -> bool

Update the last actioned view internally for votes and proposals

Returns true if the action is for a newer view than the last action of that type

pub fn reset_actions(&mut self)

reset last actions to genesis so we can resend events in tests

pub fn update_proposed_view( &mut self, proposal: Proposal<TYPES, QuorumProposalWrapper<TYPES>>, ) -> Result<()>

Update the last proposal.

Errors

Can return an error when the new view_number is not higher than the existing proposed view number.

pub fn update_last_decided_view( &mut self, view_number: TYPES::View, ) -> Result<()>

Update the last decided view.

Errors

Can return an error when the new view_number is not higher than the existing decided view number.

pub fn update_locked_view(&mut self, view_number: TYPES::View) -> Result<()>

Update the locked view.

Errors

Can return an error when the new view_number is not higher than the existing locked view number.

pub fn update_da_view( &mut self, view_number: TYPES::View, epoch: Option<TYPES::Epoch>, payload_commitment: VidCommitment, ) -> Result<()>

Update the validated state map with a new view_number/view combo.

Errors

Can return an error when the new view contains less information than the existing view with the same view number.

pub fn update_leaf( &mut self, leaf: Leaf2<TYPES>, state: Arc<TYPES::ValidatedState>, delta: Option<Arc<<TYPES::ValidatedState as ValidatedState<TYPES>>::Delta>>, ) -> Result<()>

Update the validated state map with a new view_number/view combo.

Errors

Can return an error when the new view contains less information than the existing view with the same view number.

fn update_validated_state_map( &mut self, view_number: TYPES::View, new_view: View<TYPES>, ) -> Result<()>

Update the validated state map with a new view_number/view combo.

Errors

Can return an error when the new view contains less information than the existing view with the same view number.

fn update_saved_leaves(&mut self, leaf: Leaf2<TYPES>)

Update the saved leaves with a new leaf.

pub fn update_saved_payloads( &mut self, view_number: TYPES::View, payload: Arc<TYPES::BlockPayload>, ) -> Result<()>

Update the saved payloads with a new encoded transaction.

Errors

Can return an error when there’s an existing payload corresponding to the same view number.

pub fn update_high_qc( &mut self, high_qc: QuorumCertificate2<TYPES>, ) -> Result<()>

Update the high QC if given a newer one.

Errors

Can return an error when the provided high_qc is not newer than the existing entry.

pub fn update_next_epoch_high_qc( &mut self, high_qc: NextEpochQuorumCertificate2<TYPES>, ) -> Result<()>

Update the next epoch high QC if given a newer one.

Errors

Can return an error when the provided high_qc is not newer than the existing entry.

Panics

It can’t actually panic. If the option is None, we will not call unwrap on it.

pub fn update_vid_shares( &mut self, view_number: TYPES::View, disperse: Proposal<TYPES, VidDisperseShare<TYPES>>, )

Add a new entry to the vid_shares map.

pub fn update_saved_da_certs( &mut self, view_number: TYPES::View, cert: DaCertificate2<TYPES>, )

Add a new entry to the da_certs map.

pub fn visit_leaf_ancestors<F>( &self, start_from: TYPES::View, terminator: Terminator<TYPES::View>, ok_when_finished: bool, f: F, ) -> Result<(), HotShotError<TYPES>>
where F: FnMut(&Leaf2<TYPES>, Arc<<TYPES as NodeType>::ValidatedState>, Option<Arc<<<TYPES as NodeType>::ValidatedState as ValidatedState<TYPES>>::Delta>>) -> bool,

gather information from the parent chain of leaves

Errors

If the leaf or its ancestors are not found in storage

pub fn collect_garbage( &mut self, old_anchor_view: TYPES::View, new_anchor_view: TYPES::View, )

Garbage collects based on state change right now, this removes from both the saved_payloads and validated_state_map fields of Consensus.

Panics

On inconsistent stored entries

pub fn decided_leaf(&self) -> Leaf2<TYPES>

Gets the last decided leaf.

Panics

if the last decided view’s leaf does not exist in the state map or saved leaves, which should never happen.

pub fn state( &self, view_number: TYPES::View, ) -> Option<&Arc<TYPES::ValidatedState>>

Gets the validated state with the given view number, if in the state map.

pub fn state_and_delta(&self, view_number: TYPES::View) -> StateAndDelta<TYPES>

Gets the validated state and state delta with the given view number, if in the state map.

pub fn decided_state(&self) -> Arc<TYPES::ValidatedState>

Gets the last decided validated state.

Panics

If the last decided view’s state does not exist in the state map, which should never happen.

pub async fn calculate_and_update_vid<V: Versions>( consensus: OuterConsensus<TYPES>, view: <TYPES as NodeType>::View, membership: Arc<RwLock<TYPES::Membership>>, private_key: &<TYPES::SignatureKey as SignatureKey>::PrivateKey, upgrade_lock: &UpgradeLock<TYPES, V>, ) -> Option<()>

Associated helper function: Takes LockedConsensusState which will be updated; locks it for read and write accordingly. Calculates VidDisperse based on the view, the txns and the membership, and updates vid_shares map with the signed VidDisperseShare proposals. Returned Option indicates whether the update has actually happened or not.

pub fn is_qc_forming_eqc(&self, qc: &QuorumCertificate2<TYPES>) -> bool

Return true if the QC takes part in forming an eQC, i.e. it is one of the 3-chain certificates but not the eQC itself

pub fn is_high_qc_forming_eqc(&self) -> bool

Returns true if our high qc is forming an eQC

pub fn is_leaf_forming_eqc(&self, leaf_commit: LeafCommitment<TYPES>) -> bool

Return true if the given leaf takes part in forming an eQC, i.e. it is one of the 3-chain leaves but not the eQC leaf itself

pub fn is_leaf_extended(&self, leaf_commit: LeafCommitment<TYPES>) -> bool

Returns true if the given leaf can form an extended Quorum Certificate The Extended Quorum Certificate (eQC) is the third Quorum Certificate formed in three consecutive views for the last block in the epoch.

pub fn is_leaf_for_last_block(&self, leaf_commit: LeafCommitment<TYPES>) -> bool

Returns true if a given leaf is for the last block in the epoch

pub fn is_high_qc_for_last_block(&self) -> bool

Returns true if our high QC is for the last block in the epoch

pub fn check_eqc( &self, proposed_leaf: &Leaf2<TYPES>, parent_leaf: &Leaf2<TYPES>, ) -> bool

Returns true if the parent_leaf formed an eQC for the previous epoch to the proposed_leaf