Struct hotshot_types::consensus::ConsensusUpgradableReadLockGuard

pub struct ConsensusUpgradableReadLockGuard<'a, TYPES: NodeType> {
    lock_guard: ManuallyDrop<RwLockUpgradableReadGuard<'a, Consensus<TYPES>>>,
    taken: bool,
}

A thin wrapper around RwLockUpgradableReadGuard for Consensus that leaves debug traces when the lock is freed or upgraded

Fields
lock_guard: ManuallyDrop<RwLockUpgradableReadGuard<'a, Consensus<TYPES>>>

Inner RwLockUpgradableReadGuard

taken: bool

A helper bool to indicate whether inner lock has been unsafely taken or not

Implementations

impl<'a, TYPES: NodeType> ConsensusUpgradableReadLockGuard<'a, TYPES>

pub fn new(lock_guard: RwLockUpgradableReadGuard<'a, Consensus<TYPES>>) -> Self

Creates a new instance of ConsensusUpgradableReadLockGuard with the same name as parent OuterConsensus

pub async fn upgrade(guard: Self) -> ConsensusWriteLockGuard<'a, TYPES>

Upgrades the inner RwLockUpgradableReadGuard and leaves debug traces