Module hotshot_types::constants

configurable constants for hotshot

Constants

COMBINED_NETWORK_CACHE_SIZE

the number of messages to cache in the combined network

COMBINED_NETWORK_DELAY_DURATION

the default delay duration value in milliseconds of sending on the secondary in the combined networks

COMBINED_NETWORK_MIN_PRIMARY_FAILURES

the number of messages to attempt to send over the primary network before switching to prefer the secondary network

COMBINED_NETWORK_PRIMARY_CHECK_INTERVAL

the number of messages to send over the secondary network without delay before re-attempting the (presumed down) primary network

EVENT_CHANNEL_SIZE

Default channel size for consensus event sharing

EXTERNAL_EVENT_CHANNEL_SIZE

Default channel size for HotShot -> application communication

KAD_DEFAULT_REPUB_INTERVAL_SEC

the default kademlia record republication interval (in seconds)

LOOK_AHEAD

the number of views to gather information for ahead of time

REQUEST_DATA_DELAY

The default network data request delay in milliseconds

SRS_DEGREE

For STAKE_TABLE_CAPACITY=200, the light client prover (a.k.a. hotshot-state-prover) would need to generate proof for a circuit of slightly below 2^20 gates. Thus we need to support this upperbounded degree in our Structured Reference String (SRS), the +2 is just an artifact from the jellyfish’s Plonk proof system.

UPGRADE_BEGIN_OFFSET

The offset for how far in the future the upgrade actually begins.

UPGRADE_DECIDE_BY_OFFSET

The offset for how far in the future the upgrade certificate we attach should be decided on (or else discarded).

UPGRADE_FINISH_OFFSET

The offset for how far in the future the upgrade ends.

UPGRADE_PROPOSE_OFFSET

The offset for how far in the future we will send out a QuorumProposal with an UpgradeCertificate we form. This is also how far in advance of sending a QuorumProposal we begin collecting votes on an UpgradeProposal.

WEB_SERVER_MAJOR_VERSION

Constants for WebServerNetwork and WebServer The Web CDN is not, strictly speaking, bound to the network; it can have its own versioning. Web Server CDN Version (major)

WEB_SERVER_MINOR_VERSION

Web Server CDN Version (minor)

WEB_SERVER_VERSION

Constant for Web Server CDN Version

Type Aliases

WebServerVersion

Type for Web Server CDN Version