Struct hotshot_types::data::Leaf2

pub struct Leaf2<TYPES: NodeType> {
    view_number: TYPES::View,
    justify_qc: QuorumCertificate2<TYPES>,
    next_epoch_justify_qc: Option<NextEpochQuorumCertificate2<TYPES>>,
    parent_commitment: Commitment<Self>,
    block_header: TYPES::BlockHeader,
    upgrade_certificate: Option<UpgradeCertificate<TYPES>>,
    block_payload: Option<TYPES::BlockPayload>,
    pub view_change_evidence: Option<ViewChangeEvidence2<TYPES>>,
    pub next_drb_result: Option<DrbResult>,
    pub with_epoch: bool,
}

This is the consensus-internal analogous concept to a block, and it contains the block proper, as well as the hash of its parent Leaf.

Fields
view_number: TYPES::View

CurView from leader when proposing leaf

justify_qc: QuorumCertificate2<TYPES>

Per spec, justification

next_epoch_justify_qc: Option<NextEpochQuorumCertificate2<TYPES>>

certificate that the proposal is chaining from formed by the next epoch nodes

parent_commitment: Commitment<Self>

The hash of the parent Leaf So we can ask if it extends

block_header: TYPES::BlockHeader

Block header.

upgrade_certificate: Option<UpgradeCertificate<TYPES>>

Optional upgrade certificate, if one was attached to the quorum proposal for this view.

block_payload: Option<TYPES::BlockPayload>

Optional block payload.

It may be empty for nodes not in the DA committee.

view_change_evidence: Option<ViewChangeEvidence2<TYPES>>

Possible timeout or view sync certificate. If the justify_qc is not for a proposal in the immediately preceding view, then either a timeout or view sync certificate must be attached.

next_drb_result: Option<DrbResult>

The DRB result for the next epoch.

This is required only for the last block of the epoch. Nodes will verify that it’s consistent with the result from their computations.

with_epoch: bool

Indicates whether or not epochs were enabled.

Implementations

impl<TYPES: NodeType> Leaf2<TYPES>

pub async fn genesis<V: Versions>( validated_state: &TYPES::ValidatedState, instance_state: &TYPES::InstanceState, ) -> Self

Create a new leaf from its components.

Panics

Panics if the genesis payload (TYPES::BlockPayload::genesis()) is malformed (unable to be interpreted as bytes).

pub fn view_number(&self) -> TYPES::View

Time when this leaf was created.

pub fn epoch(&self, epoch_height: u64) -> Option<TYPES::Epoch>

Epoch in which this leaf was created.

pub fn height(&self) -> u64

Height of this leaf in the chain.

Equivalently, this is the number of leaves before this one in the chain.

pub fn justify_qc(&self) -> QuorumCertificate2<TYPES>

The QC linking this leaf to its parent in the chain.

pub fn upgrade_certificate(&self) -> Option<UpgradeCertificate<TYPES>>

The QC linking this leaf to its parent in the chain.

pub fn parent_commitment(&self) -> Commitment<Self>

Commitment to this leaf’s parent.

pub fn block_header(&self) -> &<TYPES as NodeType>::BlockHeader

The block header contained in this leaf.

pub fn block_header_mut(&mut self) -> &mut <TYPES as NodeType>::BlockHeader

Get a mutable reference to the block header contained in this leaf.

pub fn fill_block_payload<V: Versions>( &mut self, block_payload: TYPES::BlockPayload, num_storage_nodes: usize, version: Version, ) -> Result<(), BlockError>

Fill this leaf with the block payload.

Errors

Fails if the payload commitment doesn’t match self.block_header.payload_commitment() or if the transactions are of invalid length

pub fn unfill_block_payload(&mut self) -> Option<TYPES::BlockPayload>

Take the block payload from the leaf and return it if it is present

pub fn fill_block_payload_unchecked( &mut self, block_payload: TYPES::BlockPayload, )

Fill this leaf with the block payload, without checking header and payload consistency

pub fn block_payload(&self) -> Option<TYPES::BlockPayload>

Optional block payload.

pub fn payload_commitment(&self) -> VidCommitment

A commitment to the block payload contained in this leaf.

pub async fn extends_upgrade( &self, parent: &Self, decided_upgrade_certificate: &Arc<RwLock<Option<UpgradeCertificate<TYPES>>>>, ) -> Result<()>

Validate that a leaf has the right upgrade certificate to be the immediate child of another leaf

This may not be a complete function. Please double-check that it performs the checks you expect before substituting validation logic with it.

Errors

Returns an error if the certificates are not identical, or that when we no longer see a cert, it’s for the right reason.

pub fn to_leaf_unsafe(self) -> Leaf<TYPES>

Converts a Leaf2 to a Leaf. This operation is fundamentally unsafe and should not be used.

impl<TYPES: NodeType> Leaf2<TYPES>

pub fn from_quorum_proposal( quorum_proposal: &QuorumProposalWrapper<TYPES>, ) -> Self

Constructs a leaf from a given quorum proposal.