Struct hotshot_types::data::ValidatingProposal

pub struct ValidatingProposal<TYPES: NodeType, LEAF>where
    LEAF: Committable + LeafType<NodeType = TYPES>,{
    pub block_commitment: Commitment<TYPES::BlockType>,
    pub view_number: TYPES::Time,
    pub height: u64,
    pub justify_qc: QuorumCertificate<TYPES, LEAF>,
    pub parent_commitment: Commitment<LEAF>,
    pub deltas: TYPES::BlockType,
    pub state_commitment: Commitment<TYPES::StateType>,
    pub rejected: Vec<<TYPES::BlockType as Block>::Transaction>,
    pub proposer_id: EncodedPublicKey,
}

subset of state that we stick into a leaf. original hotstuff proposal

Fields
block_commitment: Commitment<TYPES::BlockType>

current view’s block commitment

view_number: TYPES::Time

CurView from leader when proposing leaf

height: u64

Height from leader when proposing leaf

justify_qc: QuorumCertificate<TYPES, LEAF>

Per spec, justification

parent_commitment: Commitment<LEAF>

The hash of the parent Leaf So we can ask if it extends

deltas: TYPES::BlockType

Block leaf wants to apply

state_commitment: Commitment<TYPES::StateType>

What the state should be after applying self.deltas

rejected: Vec<<TYPES::BlockType as Block>::Transaction>

Transactions that were marked for rejection while collecting deltas

proposer_id: EncodedPublicKey

the propser id