Trait hotshot_types::data::LeafType

pub trait LeafType: Debug + Display + Clone + 'static + Committable + Serialize + for<'a> Deserialize<'a> + Send + Sync + Eq + Hash {
    type NodeType: NodeType;
    type DeltasType: DeltasType<LeafBlock<Self>>;
    type MaybeState: Clone + Debug + for<'a> Deserialize<'a> + PartialEq + Eq + Hash + Send + Serialize + Sync;

    // Required methods
    fn new(
        view_number: LeafTime<Self>,
        justify_qc: QuorumCertificate<Self::NodeType, Self>,
        deltas: LeafBlock<Self>,
        state: LeafState<Self>
    ) -> Self;
    fn get_view_number(&self) -> LeafTime<Self>;
    fn get_height(&self) -> u64;
    fn set_height(&mut self, height: u64);
    fn get_justify_qc(&self) -> QuorumCertificate<Self::NodeType, Self>;
    fn get_parent_commitment(&self) -> Commitment<Self>;
    fn get_deltas(&self) -> Self::DeltasType;
    fn fill_deltas(
        &mut self,
        block: LeafBlock<Self>
    ) -> Result<(), LeafDeltasError<Self>>;
    fn get_state(&self) -> Self::MaybeState;
    fn get_rejected(&self) -> Vec<LeafTransaction<Self>>;
    fn get_timestamp(&self) -> i128;
    fn get_proposer_id(&self) -> EncodedPublicKey;
    fn from_stored_view(stored_view: StoredView<Self::NodeType, Self>) -> Self;

    // Provided method
    fn get_deltas_commitment(&self) -> Commitment<LeafBlock<Self>> { ... }

}

An item which is appended to a blockchain.

Required Associated Types

type NodeType: NodeType

Type of nodes participating in the network.

type DeltasType: DeltasType<LeafBlock<Self>>

Type of block contained by this leaf.

type MaybeState: Clone + Debug + for<'a> Deserialize<'a> + PartialEq + Eq + Hash + Send + Serialize + Sync

Either state or empty

Required Methods

fn new( view_number: LeafTime<Self>, justify_qc: QuorumCertificate<Self::NodeType, Self>, deltas: LeafBlock<Self>, state: LeafState<Self> ) -> Self

Create a new leaf from its components.

fn get_view_number(&self) -> LeafTime<Self>

Time when this leaf was created.

fn get_height(&self) -> u64

Height of this leaf in the chain.

Equivalently, this is the number of leaves before this one in the chain.

fn set_height(&mut self, height: u64)

Change the height of this leaf.

fn get_justify_qc(&self) -> QuorumCertificate<Self::NodeType, Self>

The QC linking this leaf to its parent in the chain.

fn get_parent_commitment(&self) -> Commitment<Self>

Commitment to this leaf’s parent.

fn get_deltas(&self) -> Self::DeltasType

The block contained in this leaf.

fn fill_deltas( &mut self, block: LeafBlock<Self> ) -> Result<(), LeafDeltasError<Self>>

Fill this leaf with the entire corresponding block.

After this function succeeds, self.get_deltas().try_resolve() is guaranteed to return Ok(block).

Errors

Fails if block does not match self.get_deltas_commitment(), or if the block is not able to be stored for some implementation-defined reason.

fn get_state(&self) -> Self::MaybeState

The blockchain state after appending this leaf.

fn get_rejected(&self) -> Vec<LeafTransaction<Self>>

Transactions rejected or invalidated by the application of this leaf.

fn get_timestamp(&self) -> i128

Real-world time when this leaf was created.

fn get_proposer_id(&self) -> EncodedPublicKey

Identity of the network participant who proposed this leaf.

fn from_stored_view(stored_view: StoredView<Self::NodeType, Self>) -> Self

Create a leaf from information stored about a view.

Provided Methods

fn get_deltas_commitment(&self) -> Commitment<LeafBlock<Self>>

A commitment to the block contained in this leaf.