hotshot_types::data

Module vid_disperse

This module provides types for VID disperse related data structures.

Structs

ADVZDisperse

ADVZ dispersal data

ADVZDisperseShare

ADVZ share and associated metadata for a single node

AvidMDisperse

ADVZ dispersal data

VidDisperseShare2

VID share and associated metadata for a single node

Weights 🔒

The weights and total weight used in VID calculations

Constants

VID_TARGET_TOTAL_STAKE

The target total stake to scale to for VID.

Functions

approximate_weights 🔒

vid_total_weight