Module hotshot_types::light_client

Types and structs associated with light client state

Structs

GenericLightClientState

A light client state

GenericPublicInput

Public input to the light client state prover service

GenericStakeTableState

Stake table state

StateKeyPair

Key pairs for signing/verifying a light client state

StateSignatureRequestBody

Request body to send to the state relay server

StateSignaturesBundle

The state signatures bundle is a light client state and its signatures collected

Type Aliases

CircuitField

Base field in the prover circuit

LightClientState

Concrete type for light client state

PublicInput

Concrete for circuit’s public input

StakeTableState

Concrete type for stake table state

StateSignKey

Signing key for signing a light client state

StateSignature

Signatures

StateSignatureScheme

Signature scheme

StateVerKey

Verification key for verifying state signatures