Struct hotshot_types::network::NetworkConfig

pub struct NetworkConfig<KEY: SignatureKey> {

    pub rounds: usize,
    pub indexed_da: bool,
    pub transactions_per_round: usize,
    pub manual_start_password: Option<String>,
    pub num_bootrap: usize,
    pub next_view_timeout: u64,
    pub view_sync_timeout: Duration,
    pub builder_timeout: Duration,
    pub data_request_delay: Duration,
    pub node_index: u64,
    pub seed: [u8; 32],
    pub transaction_size: usize,
    pub key_type_name: String,
    pub libp2p_config: Option<Libp2pConfig>,
    pub config: HotShotConfig<KEY>,
    pub cdn_marshal_address: Option<String>,
    pub combined_network_config: Option<CombinedNetworkConfig>,
    pub commit_sha: String,
    pub builder: BuilderType,
    pub random_builder: Option<RandomBuilderConfig>,
    pub public_keys: Vec<PeerConfigKeys<KEY>>,

}

a network configuration

Fields
rounds: usize

number of views to run

indexed_da: bool

whether DA membership is determined by index. if true, the first k nodes to register form the DA committee if false, DA membership is requested by the nodes

transactions_per_round: usize

number of transactions per view

manual_start_password: Option<String>

password to have the orchestrator start the network, regardless of the number of nodes connected.

num_bootrap: usize

number of bootstrap nodes

next_view_timeout: u64

timeout before starting the next view

view_sync_timeout: Duration

timeout before starting next view sync round

builder_timeout: Duration

The maximum amount of time a leader can wait to get a block from a builder

data_request_delay: Duration

time to wait until we request data associated with a proposal

node_index: u64

global index of node (for testing purposes a uid)

seed: [u8; 32]

unique seed (for randomness? TODO)

transaction_size: usize

size of transactions

key_type_name: String

name of the key type (for debugging)

libp2p_config: Option<Libp2pConfig>

the libp2p config

config: HotShotConfig<KEY>

the hotshot config

cdn_marshal_address: Option<String>

The address for the Push CDNโs โmarshalโ, A.K.A. load balancer

combined_network_config: Option<CombinedNetworkConfig>

combined network config

commit_sha: String

the commit this run is based on

builder: BuilderType

builder to use

random_builder: Option<RandomBuilderConfig>

random builder config

public_keys: Vec<PeerConfigKeys<KEY>>

The list of public keys that are allowed to connect to the orchestrator

Implementations

impl<K: SignatureKey> NetworkConfig<K>

pub fn generate_init_validator_config( cur_node_index: u16, is_da: bool, ) -> ValidatorConfig<K>

Get a temporary node index for generating a validator config

pub fn from_file(file: String) -> Result<Self, NetworkConfigError>

Loads a NetworkConfig from a file.

This function takes a file path as a string, reads the file, and then deserializes the contents into a NetworkConfig.

Arguments

file - A string representing the path to the file from which to load the NetworkConfig.

Returns

This function returns a Result that contains a NetworkConfig if the file was successfully read and deserialized, or a NetworkConfigError if an error occurred.

Errors

This function will return an error if the file cannot be read or if the contents cannot be deserialized into a NetworkConfig.

Examples

// # use hotshot::traits::election::static_committee::StaticElectionConfig;
let file = "/path/to/my/config".to_string();
// NOTE: broken due to staticelectionconfig not being importable
// cannot import staticelectionconfig from hotshot without creating circular dependency
// making this work probably involves the `types` crate implementing a dummy
// electionconfigtype just to make this example work
let config = NetworkConfig::<BLSPubKey, StaticElectionConfig>::from_file(file).unwrap();

pub fn to_file(&self, file: String) -> Result<(), NetworkConfigError>

Serializes the NetworkConfig and writes it to a file.

This function takes a file path as a string, serializes the NetworkConfig into JSON format using serde_json and then writes the serialized data to the file.

Arguments

file - A string representing the path to the file where the NetworkConfig should be saved.

Returns

This function returns a Result that contains () if the NetworkConfig was successfully serialized and written to the file, or a NetworkConfigError if an error occurred.

Errors

This function will return an error if the NetworkConfig cannot be serialized or if the file cannot be written.

Examples

let file = "/path/to/my/config".to_string();
let config = NetworkConfig::from_file(file);
config.to_file(file).unwrap();