Module hotshot_types::simple_certificate

Implementations of the simple certificate type. Used for Quorum, DA, and Timeout Certificates

Structs

OneHonestThreshold

Defines a threshold which is f + 1 (i.e at least one of the stake is honest)

SimpleCertificate

A certificate which can be created by aggregating many simple votes on the commitment.

SuccessThreshold

Defines a threshold which is 2f + 1 (Amount needed for Quorum)

UpgradeThreshold

Defines a threshold which is 0.9n + 1 (i.e. over 90% of the nodes with stake)

Traits

Threshold

Trait which allows use to inject different threshold calculations into a Certificate type

Type Aliases

DACertificate

Type alias for a DA certificate over DAData

QuorumCertificate

Type alias for a QuorumCertificate, which is a SimpleCertificate of QuorumVotes

TimeoutCertificate

Type alias for a Timeout certificate over a view number

UpgradeCertificate

Type alias for a UpgradeCertificate, which is a SimpleCertificate of UpgradeProposalData

ViewSyncCommitCertificate2

Type alias for a ViewSyncCommit certificate over a view number

ViewSyncFinalizeCertificate2

Type alias for a ViewSyncFinalize certificate over a view number

ViewSyncPreCommitCertificate2

Type alias for a ViewSyncPreCommit certificate over a view number