Struct hotshot_types::simple_certificate::SimpleCertificate

pub struct SimpleCertificate<TYPES: NodeType, VOTEABLE: Voteable<TYPES>, THRESHOLD: Threshold<TYPES>> {
    pub data: VOTEABLE,
    vote_commitment: Commitment<VOTEABLE>,
    pub view_number: TYPES::View,
    pub signatures: Option<<TYPES::SignatureKey as SignatureKey>::QcType>,
    pub _pd: PhantomData<(TYPES, THRESHOLD)>,
}

A certificate which can be created by aggregating many simple votes on the commitment.

Fields
data: VOTEABLE

The data this certificate is for. I.e the thing that was voted on to create this Certificate

vote_commitment: Commitment<VOTEABLE>

commitment of all the votes this cert should be signed over

view_number: TYPES::View

Which view this QC relates to

signatures: Option<<TYPES::SignatureKey as SignatureKey>::QcType>

assembled signature for certificate aggregation

_pd: PhantomData<(TYPES, THRESHOLD)>

phantom data for THRESHOLD and TYPES

Implementations

impl<TYPES: NodeType> SimpleCertificate<TYPES, QuorumData<TYPES>, SuccessThreshold>

pub async fn genesis<V: Versions>( validated_state: &TYPES::ValidatedState, instance_state: &TYPES::InstanceState, ) -> Self

Creat the Genesis certificate

impl<TYPES: NodeType> SimpleCertificate<TYPES, QuorumData2<TYPES>, SuccessThreshold>

pub async fn genesis<V: Versions>( validated_state: &TYPES::ValidatedState, instance_state: &TYPES::InstanceState, ) -> Self

Create the Genesis certificate

impl<TYPES: NodeType, VOTEABLE: Voteable<TYPES>, THRESHOLD: Threshold<TYPES>> SimpleCertificate<TYPES, VOTEABLE, THRESHOLD>

pub fn new( data: VOTEABLE, vote_commitment: Commitment<VOTEABLE>, view_number: TYPES::View, signatures: Option<<TYPES::SignatureKey as SignatureKey>::QcType>, pd: PhantomData<(TYPES, THRESHOLD)>, ) -> Self

Creates a new instance of SimpleCertificate

impl<TYPES: NodeType> SimpleCertificate<TYPES, UpgradeProposalData<TYPES>, UpgradeThreshold>

pub async fn is_relevant( &self, view_number: TYPES::View, decided_upgrade_certificate: Arc<RwLock<Option<Self>>>, ) -> Result<()>

Determines whether or not a certificate is relevant (i.e. we still have time to reach a decide)

Errors

Returns an error when the certificate is no longer relevant

pub async fn validate<V: Versions>( upgrade_certificate: &Option<Self>, membership: &RwLock<TYPES::Membership>, epoch: Option<TYPES::Epoch>, upgrade_lock: &UpgradeLock<TYPES, V>, ) -> Result<()>

Validate an upgrade certificate.

Errors

Returns an error when the upgrade certificate is invalid.

pub fn upgrading_in(&self, view: TYPES::View) -> bool

Given an upgrade certificate and a view, tests whether the view is in the period where we are upgrading, which requires that we propose with null blocks.

impl<TYPES: NodeType> SimpleCertificate<TYPES, QuorumData<TYPES>, SuccessThreshold>

pub fn to_qc2(self) -> QuorumCertificate2<TYPES>

Convert a QuorumCertificate into a QuorumCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, QuorumData2<TYPES>, SuccessThreshold>

pub fn to_qc(self) -> QuorumCertificate<TYPES>

Convert a QuorumCertificate2 into a QuorumCertificate

impl<TYPES: NodeType> SimpleCertificate<TYPES, DaData, SuccessThreshold>

pub fn to_dac2(self) -> DaCertificate2<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, DaData2<TYPES>, SuccessThreshold>

pub fn to_dac(self) -> DaCertificate<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, ViewSyncPreCommitData<TYPES>, OneHonestThreshold>

pub fn to_vsc2(self) -> ViewSyncPreCommitCertificate2<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, ViewSyncPreCommitData2<TYPES>, OneHonestThreshold>

pub fn to_vsc(self) -> ViewSyncPreCommitCertificate<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, ViewSyncCommitData<TYPES>, SuccessThreshold>

pub fn to_vsc2(self) -> ViewSyncCommitCertificate2<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, ViewSyncCommitData2<TYPES>, SuccessThreshold>

pub fn to_vsc(self) -> ViewSyncCommitCertificate<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, ViewSyncFinalizeData<TYPES>, SuccessThreshold>

pub fn to_vsc2(self) -> ViewSyncFinalizeCertificate2<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, ViewSyncFinalizeData2<TYPES>, SuccessThreshold>

pub fn to_vsc(self) -> ViewSyncFinalizeCertificate<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, TimeoutData<TYPES>, SuccessThreshold>

pub fn to_tc2(self) -> TimeoutCertificate2<TYPES>

Convert a DaCertificate into a DaCertificate2

impl<TYPES: NodeType> SimpleCertificate<TYPES, TimeoutData2<TYPES>, SuccessThreshold>

pub fn to_tc(self) -> TimeoutCertificate<TYPES>

Convert a DaCertificate into a DaCertificate2