Type Alias hotshot_types::simple_certificate::QuorumCertificate

pub type QuorumCertificate<TYPES> = SimpleCertificate<TYPES, QuorumData<TYPES>, SuccessThreshold>;

Type alias for a QuorumCertificate, which is a SimpleCertificate over QuorumData

Aliased Type

struct QuorumCertificate<TYPES> {
    pub data: QuorumData<TYPES>,
    vote_commitment: Commitment<QuorumData<TYPES>>,
    pub view_number: <TYPES as NodeType>::View,
    pub signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QcType>,
    pub _pd: PhantomData<(TYPES, SuccessThreshold)>,
}

Fields
data: QuorumData<TYPES>

The data this certificate is for. I.e the thing that was voted on to create this Certificate

vote_commitment: Commitment<QuorumData<TYPES>>

commitment of all the votes this cert should be signed over

view_number: <TYPES as NodeType>::View

Which view this QC relates to

signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QcType>

assembled signature for certificate aggregation

_pd: PhantomData<(TYPES, SuccessThreshold)>

phantom data for THRESHOLD and TYPES

Implementations

impl<TYPES: NodeType> QuorumCertificate<TYPES>

pub async fn genesis<V: Versions>( validated_state: &TYPES::ValidatedState, instance_state: &TYPES::InstanceState, ) -> Self

Creat the Genesis certificate

impl<TYPES: NodeType> QuorumCertificate<TYPES>

pub fn to_qc2(self) -> QuorumCertificate2<TYPES>

Convert a QuorumCertificate into a QuorumCertificate2