hotshot_types::simple_certificate
Type Alias QuorumCertificate2

pub type QuorumCertificate2<TYPES> = SimpleCertificate<TYPES, QuorumData2<TYPES>, SuccessThreshold>;

Type alias for a QuorumCertificate2, which is a SimpleCertificate over QuorumData2

Aliased Type

struct QuorumCertificate2<TYPES> {
    pub data: QuorumData2<TYPES>,
    vote_commitment: Commitment<QuorumData2<TYPES>>,
    pub view_number: <TYPES as NodeType>::View,
    pub signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QcType>,
    pub _pd: PhantomData<(TYPES, SuccessThreshold)>,
}

Fields
data: QuorumData2<TYPES>

The data this certificate is for. I.e the thing that was voted on to create this Certificate

vote_commitment: Commitment<QuorumData2<TYPES>>

commitment of all the votes this cert should be signed over

view_number: <TYPES as NodeType>::View

Which view this QC relates to

signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QcType>

assembled signature for certificate aggregation

_pd: PhantomData<(TYPES, SuccessThreshold)>

phantom data for THRESHOLD and TYPES

Implementations

impl<TYPES: NodeType> QuorumCertificate2<TYPES>

pub async fn genesis<V: Versions>( validated_state: &TYPES::ValidatedState, instance_state: &TYPES::InstanceState, ) -> Self

Create the Genesis certificate

impl<TYPES: NodeType> QuorumCertificate2<TYPES>

pub fn to_qc(self) -> QuorumCertificate<TYPES>

Convert a QuorumCertificate2 into a QuorumCertificate