Type Alias hotshot_types::simple_certificate::TimeoutCertificate

pub type TimeoutCertificate<TYPES> = SimpleCertificate<TYPES, TimeoutData<TYPES>, SuccessThreshold>;

Type alias for a Timeout certificate over a view number

Aliased Type

struct TimeoutCertificate<TYPES> {
    pub data: TimeoutData<TYPES>,
    pub vote_commitment: Commitment<TimeoutData<TYPES>>,
    pub view_number: <TYPES as NodeType>::Time,
    pub signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QCType>,
    pub is_genesis: bool,
    pub _pd: PhantomData<(TYPES, SuccessThreshold)>,
}

Fields
data: TimeoutData<TYPES>

The data this certificate is for. I.e the thing that was voted on to create this Certificate

vote_commitment: Commitment<TimeoutData<TYPES>>

commitment of all the votes this cert should be signed over

view_number: <TYPES as NodeType>::Time

Which view this QC relates to

signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QCType>

assembled signature for certificate aggregation

is_genesis: bool

If this QC is for the genesis block

_pd: PhantomData<(TYPES, SuccessThreshold)>

phantom data for THRESHOLD and TYPES