Type Alias hotshot_types::simple_certificate::ViewSyncCommitCertificate2

pub type ViewSyncCommitCertificate2<TYPES> = SimpleCertificate<TYPES, ViewSyncCommitData<TYPES>, SuccessThreshold>;

Type alias for a ViewSyncCommit certificate over a view number

Aliased Type

struct ViewSyncCommitCertificate2<TYPES> {
    pub data: ViewSyncCommitData<TYPES>,
    pub vote_commitment: Commitment<ViewSyncCommitData<TYPES>>,
    pub view_number: <TYPES as NodeType>::Time,
    pub signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QCType>,
    pub is_genesis: bool,
    pub _pd: PhantomData<(TYPES, SuccessThreshold)>,
}

Fields
data: ViewSyncCommitData<TYPES>

The data this certificate is for. I.e the thing that was voted on to create this Certificate

vote_commitment: Commitment<ViewSyncCommitData<TYPES>>

commitment of all the votes this cert should be signed over

view_number: <TYPES as NodeType>::Time

Which view this QC relates to

signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QCType>

assembled signature for certificate aggregation

is_genesis: bool

If this QC is for the genesis block

_pd: PhantomData<(TYPES, SuccessThreshold)>

phantom data for THRESHOLD and TYPES