Type Alias hotshot_types::simple_certificate::ViewSyncPreCommitCertificate2

pub type ViewSyncPreCommitCertificate2<TYPES> = SimpleCertificate<TYPES, ViewSyncPreCommitData<TYPES>, OneHonestThreshold>;

Type alias for a ViewSyncPreCommit certificate over a view number

Aliased Type

struct ViewSyncPreCommitCertificate2<TYPES> {
    pub data: ViewSyncPreCommitData<TYPES>,
    pub vote_commitment: Commitment<ViewSyncPreCommitData<TYPES>>,
    pub view_number: <TYPES as NodeType>::Time,
    pub signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QCType>,
    pub is_genesis: bool,
    pub _pd: PhantomData<(TYPES, OneHonestThreshold)>,
}

Fields
data: ViewSyncPreCommitData<TYPES>

The data this certificate is for. I.e the thing that was voted on to create this Certificate

vote_commitment: Commitment<ViewSyncPreCommitData<TYPES>>

commitment of all the votes this cert should be signed over

view_number: <TYPES as NodeType>::Time

Which view this QC relates to

signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QCType>

assembled signature for certificate aggregation

is_genesis: bool

If this QC is for the genesis block

_pd: PhantomData<(TYPES, OneHonestThreshold)>

phantom data for THRESHOLD and TYPES