hotshot_types::simple_vote
Trait Voteable

pub trait Voteable<TYPES: NodeType>:
    Sealed
    + Committable
    + Clone
    + Serialize
    + Debug
    + PartialEq
    + Hash
    + Eq { }

Marker trait for data or commitments that can be voted on. Only structs in this file can implement voteable. This is enforced with the Sealed trait Sealing this trait prevents creating new vote types outside this file.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.