Module hotshot_types::traits

Common traits for the HotShot protocol

Re-exports

pub use block_contents::Block;

pub use state::State;

Modules

block_contents

Abstraction over the contents of a block

election

The election trait, used to decide which node is the leader and determine if a vote is valid.

metrics

The Metrics trait is used to collect information from multiple components in the entire system.

network

Network access compatibility

node_implementation

Composite trait for node behavior

signature_key

Minimal compatibility over public key signatures

state

Abstraction over the global state that blocks modify

storage

Abstraction over on-disk storage of node state