Trait hotshot_types::traits::block_contents::BlockHeader

pub trait BlockHeader<TYPES: NodeType>: Serialize + Clone + Debug + Hash + PartialEq + Eq + Send + Sync + DeserializeOwned + Committable {
    // Required methods
    fn new(
        parent_state: &TYPES::ValidatedState,
        instance_state: &<TYPES::ValidatedState as ValidatedState<TYPES>>::Instance,
        parent_leaf: &Leaf<TYPES>,
        payload_commitment: VidCommitment,
        metadata: <TYPES::BlockPayload as BlockPayload>::Metadata,
        builder_fee: BuilderFee<TYPES>
    ) -> impl Future<Output = Self> + Send;
    fn genesis(
        instance_state: &<TYPES::ValidatedState as ValidatedState<TYPES>>::Instance,
        payload_commitment: VidCommitment,
        metadata: <TYPES::BlockPayload as BlockPayload>::Metadata
    ) -> Self;
    fn block_number(&self) -> u64;
    fn payload_commitment(&self) -> VidCommitment;
    fn metadata(&self) -> &<TYPES::BlockPayload as BlockPayload>::Metadata;
    fn builder_commitment(
        &self,
        metadata: &<TYPES::BlockPayload as BlockPayload>::Metadata
    ) -> BuilderCommitment;
}

Header of a block, which commits to a BlockPayload.

Required Methods

fn new( parent_state: &TYPES::ValidatedState, instance_state: &<TYPES::ValidatedState as ValidatedState<TYPES>>::Instance, parent_leaf: &Leaf<TYPES>, payload_commitment: VidCommitment, metadata: <TYPES::BlockPayload as BlockPayload>::Metadata, builder_fee: BuilderFee<TYPES> ) -> impl Future<Output = Self> + Send

Build a header with the parent validate state, instance-level state, parent leaf, payload commitment, and metadata.

fn genesis( instance_state: &<TYPES::ValidatedState as ValidatedState<TYPES>>::Instance, payload_commitment: VidCommitment, metadata: <TYPES::BlockPayload as BlockPayload>::Metadata ) -> Self

Build the genesis header, payload, and metadata.

fn block_number(&self) -> u64

Get the block number.

fn payload_commitment(&self) -> VidCommitment

Get the payload commitment.

fn metadata(&self) -> &<TYPES::BlockPayload as BlockPayload>::Metadata

Get the metadata.

fn builder_commitment( &self, metadata: &<TYPES::BlockPayload as BlockPayload>::Metadata ) -> BuilderCommitment

Get the builder commitment

Object Safety

This trait is not object safe.