Trait hotshot_types::traits::block_contents::BlockPayload

pub trait BlockPayload: Serialize + Clone + Debug + Display + Hash + PartialEq + Eq + Send + Sync + DeserializeOwned {
    type Error: Error + Debug + Send + Sync;
    type Transaction: Transaction;
    type Metadata: Clone + Debug + DeserializeOwned + Eq + Hash + Send + Sync + Serialize;
    type Encode<'a>: 'a + Iterator<Item = u8> + Send
       where Self: 'a;

    // Required methods
    fn from_transactions(
        transactions: impl IntoIterator<Item = Self::Transaction>
    ) -> Result<(Self, Self::Metadata), Self::Error>;
    fn from_bytes<I>(encoded_transactions: I, metadata: &Self::Metadata) -> Self
       where I: Iterator<Item = u8>;
    fn genesis() -> (Self, Self::Metadata);
    fn encode(&self) -> Result<Self::Encode<'_>, Self::Error>;
    fn transaction_commitments(
        &self,
        metadata: &Self::Metadata
    ) -> Vec<Commitment<Self::Transaction>>;
}

Abstraction over the full contents of a block

This trait encapsulates the behaviors that the transactions of a block must have in order to be used by consensus

Must have a predefined error type (BlockPayload::Error)

Must have a transaction type that can be compared for equality, serialized and serialized, sent between threads, and can have a hash produced of it

Must be hashable

Required Associated Types

type Error: Error + Debug + Send + Sync

The error type for this type of block

type Transaction: Transaction

The type of the transitions we are applying

type Metadata: Clone + Debug + DeserializeOwned + Eq + Hash + Send + Sync + Serialize

Data created during block building which feeds into the block header

type Encode<'a>: 'a + Iterator<Item = u8> + Send where Self: 'a

Encoded payload.

Required Methods

fn from_transactions( transactions: impl IntoIterator<Item = Self::Transaction> ) -> Result<(Self, Self::Metadata), Self::Error>

Build a payload and associated metadata with the transactions.

Errors

If the transaction length conversion fails.

fn from_bytes<I>(encoded_transactions: I, metadata: &Self::Metadata) -> Selfwhere I: Iterator<Item = u8>,

Build a payload with the encoded transaction bytes, metadata, and the associated number of VID storage nodes

I may be, but not necessarily is, the Encode type directly from fn encode.

fn genesis() -> (Self, Self::Metadata)

Build the genesis payload and metadata.

fn encode(&self) -> Result<Self::Encode<'_>, Self::Error>

Encode the payload

Errors

If the transaction length conversion fails.

fn transaction_commitments( &self, metadata: &Self::Metadata ) -> Vec<Commitment<Self::Transaction>>

List of transaction commitments.

Object Safety

This trait is not object safe.