Trait hotshot_types::traits::election::ConsensusExchange

pub trait ConsensusExchange<TYPES: NodeType, M: NetworkMsg>: Send + Sync {
    type Proposal: ProposalType<NodeType = TYPES>;
    type Vote: VoteType<TYPES>;
    type Certificate: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, Self::Commitment> + Hash + Eq;
    type Membership: Membership<TYPES>;
    type Networking: CommunicationChannel<TYPES, M, Self::Proposal, Self::Vote, Self::Membership>;
    type Commitment: Committable + Serialize + Clone;

    // Required methods
    fn create(
        keys: Vec<TYPES::SignatureKey>,
        config: TYPES::ElectionConfigType,
        network: Self::Networking,
        pk: TYPES::SignatureKey,
        sk: <TYPES::SignatureKey as SignatureKey>::PrivateKey,
        ek: KeyPair
    ) -> Self;
    fn network(&self) -> &Self::Networking;
    fn vote_data(
        &self,
        commit: Commitment<Self::Commitment>
    ) -> VoteData<Self::Commitment>;
    fn accumulate_vote(
        &self,
        encoded_key: &EncodedPublicKey,
        encoded_signature: &EncodedSignature,
        leaf_commitment: Commitment<Self::Commitment>,
        vote_data: VoteData<Self::Commitment>,
        vote_token: TYPES::VoteTokenType,
        view_number: TYPES::Time,
        accumlator: VoteAccumulator<TYPES::VoteTokenType, Self::Commitment>
    ) -> Either<VoteAccumulator<TYPES::VoteTokenType, Self::Commitment>, Self::Certificate>;
    fn membership(&self) -> &Self::Membership;
    fn public_key(&self) -> &TYPES::SignatureKey;
    fn private_key(&self) -> &<TYPES::SignatureKey as SignatureKey>::PrivateKey;

    // Provided methods
    fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey { ... }
    fn is_leader(&self, view_number: TYPES::Time) -> bool { ... }
    fn success_threshold(&self) -> NonZeroU64 { ... }
    fn failure_threshold(&self) -> NonZeroU64 { ... }
    fn make_vote_token(
        &self,
        view_number: TYPES::Time
    ) -> Result<Option<TYPES::VoteTokenType>, ElectionError> { ... }
    fn is_valid_cert(
        &self,
        qc: &Self::Certificate,
        commit: Commitment<Self::Commitment>
    ) -> bool { ... }
    fn is_valid_vote(
        &self,
        encoded_key: &EncodedPublicKey,
        encoded_signature: &EncodedSignature,
        data: VoteData<Self::Commitment>,
        view_number: TYPES::Time,
        vote_token: Checked<TYPES::VoteTokenType>
    ) -> bool { ... }

}

Protocol for exchanging proposals and votes to make decisions in a distributed network.

An instance of ConsensusExchange represents the state of one participant in the protocol, allowing them to vote and query information about the overall state of the protocol (such as membership and leader status).

Required Associated Types

type Proposal: ProposalType<NodeType = TYPES>

A proposal for participants to vote on.

type Vote: VoteType<TYPES>

A vote on a Proposal.

type Certificate: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, Self::Commitment> + Hash + Eq

A SignedCertificate attesting to a decision taken by the committee.

type Membership: Membership<TYPES>

The committee eligible to make decisions.

type Networking: CommunicationChannel<TYPES, M, Self::Proposal, Self::Vote, Self::Membership>

Network used by Membership to communicate.

type Commitment: Committable + Serialize + Clone

Commitments to items which are the subject of proposals and decisions.

Required Methods

fn create( keys: Vec<TYPES::SignatureKey>, config: TYPES::ElectionConfigType, network: Self::Networking, pk: TYPES::SignatureKey, sk: <TYPES::SignatureKey as SignatureKey>::PrivateKey, ek: KeyPair ) -> Self

Join a ConsensusExchange with the given identity (pk and sk).

fn network(&self) -> &Self::Networking

The network being used by this exchange.

fn vote_data( &self, commit: Commitment<Self::Commitment> ) -> VoteData<Self::Commitment>

The contents of a vote on commit.

fn accumulate_vote( &self, encoded_key: &EncodedPublicKey, encoded_signature: &EncodedSignature, leaf_commitment: Commitment<Self::Commitment>, vote_data: VoteData<Self::Commitment>, vote_token: TYPES::VoteTokenType, view_number: TYPES::Time, accumlator: VoteAccumulator<TYPES::VoteTokenType, Self::Commitment> ) -> Either<VoteAccumulator<TYPES::VoteTokenType, Self::Commitment>, Self::Certificate>

Add a vote to the accumulating signature. Return The certificate if the vote brings us over the threshould, Else return the accumulator.

fn membership(&self) -> &Self::Membership

The committee which votes on proposals.

fn public_key(&self) -> &TYPES::SignatureKey

This participant’s public key.

fn private_key(&self) -> &<TYPES::SignatureKey as SignatureKey>::PrivateKey

This participant’s private key.

Provided Methods

fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey

The leader of the Membership at time view_number.

fn is_leader(&self, view_number: TYPES::Time) -> bool

Whether this participant is leader at time view_number.

fn success_threshold(&self) -> NonZeroU64

Threshold required to approve a Proposal.

fn failure_threshold(&self) -> NonZeroU64

Threshold required to know a success threshold will not be reached

fn make_vote_token( &self, view_number: TYPES::Time ) -> Result<Option<TYPES::VoteTokenType>, ElectionError>

Attempts to generate a vote token for participation at time view_number.

Errors

When unable to make a vote token because not part of the committee

fn is_valid_cert( &self, qc: &Self::Certificate, commit: Commitment<Self::Commitment> ) -> bool

Validate a QC.

fn is_valid_vote( &self, encoded_key: &EncodedPublicKey, encoded_signature: &EncodedSignature, data: VoteData<Self::Commitment>, view_number: TYPES::Time, vote_token: Checked<TYPES::VoteTokenType> ) -> bool

Validate a vote by checking its signature and token.