Trait hotshot_types::traits::election::Election

pub trait Election<TYPES: NodeType>: Clone + Eq + PartialEq + Send + Sync + 'static {
    type StakeTable: Send + Sync;
    type QuorumCertificate: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, Self::LeafType> + Clone + Debug + Eq + Hash + PartialEq;
    type DACertificate: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, TYPES::BlockType> + Clone + Debug + Eq + PartialEq;
    type LeafType: LeafType<NodeType = TYPES, QuorumCertificate = Self::QuorumCertificate>;

    fn default_election_config(num_nodes: u64) -> TYPES::ElectionConfigType;
    fn create_election(
        keys: Vec<TYPES::SignatureKey>,
        config: TYPES::ElectionConfigType
    ) -> Self;
    fn get_stake_table(
        &self,
        view_number: TYPES::Time,
        state: &TYPES::StateType
    ) -> Self::StakeTable;
    fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey;
    fn get_committee(
        &self,
        view_number: TYPES::Time
    ) -> BTreeSet<TYPES::SignatureKey>;
    fn make_vote_token(
        &self,
        view_number: TYPES::Time,
        priv_key: &<TYPES::SignatureKey as SignatureKey>::PrivateKey
    ) -> Result<Option<TYPES::VoteTokenType>, ElectionError>;
    fn validate_vote_token(
        &self,
        view_number: TYPES::Time,
        pub_key: TYPES::SignatureKey,
        token: Checked<TYPES::VoteTokenType>
    ) -> Result<Checked<TYPES::VoteTokenType>, ElectionError>;
    fn threshold(&self) -> NonZeroU64;

    fn is_valid_dac(
        &self,
        dac: &<Self::LeafType as LeafType>::DACertificate,
        block_commitment: Commitment<TYPES::BlockType>
    ) -> bool { ... }
    fn is_valid_qc(
        &self,
        qc: &<Self::LeafType as LeafType>::QuorumCertificate
    ) -> bool { ... }
    fn is_valid_vote(
        &self,
        encoded_key: &EncodedPublicKey,
        encoded_signature: &EncodedSignature,
        data: VoteData<TYPES, Self::LeafType>,
        view_number: TYPES::Time,
        vote_token: Checked<TYPES::VoteTokenType>
    ) -> bool { ... }
    fn accumulate_vote<C: Committable, Cert>(
        &self,
        vota_meta: VoteMetaData<TYPES, C, TYPES::VoteTokenType, TYPES::Time, Self::LeafType>,
        accumulator: CertificateAccumulator<TYPES::VoteTokenType, C>
    ) -> Either<CertificateAccumulator<TYPES::VoteTokenType, C>, Cert>
    where
        Cert: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, C>,
    { ... }
}

Describes how HotShot chooses committees and leaders TODO (da) make a separate vote token type for DA and QC @ny thinks we should make the vote token types be bound to ConsensusType

Required Associated Types

type StakeTable: Send + Sync

Data structure describing the currently valid states TODO make this a trait so we can pass in places

type QuorumCertificate: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, Self::LeafType> + Clone + Debug + Eq + Hash + PartialEq

certificate for quorum on consenus

type DACertificate: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, TYPES::BlockType> + Clone + Debug + Eq + PartialEq

certificate for data availability

type LeafType: LeafType<NodeType = TYPES, QuorumCertificate = Self::QuorumCertificate>

Required Methods

fn default_election_config(num_nodes: u64) -> TYPES::ElectionConfigType

generate a default election configuration

fn create_election(
keys: Vec<TYPES::SignatureKey>,
config: TYPES::ElectionConfigType
) -> Self

create an election TODO may want to move this to a testableelection trait

fn get_stake_table(
&self,
view_number: TYPES::Time,
state: &TYPES::StateType
) -> Self::StakeTable

Returns the table from the current committed state

fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey

Returns leader for the current view number, given the current stake table

fn get_committee(
&self,
view_number: TYPES::Time
) -> BTreeSet<TYPES::SignatureKey>

fn make_vote_token(
&self,
view_number: TYPES::Time,
priv_key: &<TYPES::SignatureKey as SignatureKey>::PrivateKey
) -> Result<Option<TYPES::VoteTokenType>, ElectionError>

Attempts to generate a vote token for self

Returns None if the number of seats would be zero

Errors

TODO tbd

fn validate_vote_token(
&self,
view_number: TYPES::Time,
pub_key: TYPES::SignatureKey,
token: Checked<TYPES::VoteTokenType>
) -> Result<Checked<TYPES::VoteTokenType>, ElectionError>

Checks the claims of a received vote token

Errors

TODO tbd

fn threshold(&self) -> NonZeroU64

Returns the threshold for a specific Election implementation

Provided Methods

fn is_valid_dac(
&self,
dac: &<Self::LeafType as LeafType>::DACertificate,
block_commitment: Commitment<TYPES::BlockType>
) -> bool

Validate a DAC by checking its votes.

fn is_valid_qc(
&self,
qc: &<Self::LeafType as LeafType>::QuorumCertificate
) -> bool

Validate a QC by checking its votes.

fn is_valid_vote(
&self,
encoded_key: &EncodedPublicKey,
encoded_signature: &EncodedSignature,
data: VoteData<TYPES, Self::LeafType>,
view_number: TYPES::Time,
vote_token: Checked<TYPES::VoteTokenType>
) -> bool

Validate a vote by checking its signature and token.

fn accumulate_vote<C: Committable, Cert>(
&self,
vota_meta: VoteMetaData<TYPES, C, TYPES::VoteTokenType, TYPES::Time, Self::LeafType>,
accumulator: CertificateAccumulator<TYPES::VoteTokenType, C>
) -> Either<CertificateAccumulator<TYPES::VoteTokenType, C>, Cert>where
Cert: SignedCertificate<TYPES::SignatureKey, TYPES::Time, TYPES::VoteTokenType, C>,