Trait hotshot_types::traits::election::SignedCertificate

pub trait SignedCertificate<TYPES: NodeType, TIME, TOKEN, COMMITMENT>where
    Self: Send + Sync + Clone + Serialize + for<'a> Deserialize<'a>,
    COMMITMENT: CommitmentBounds,
    TOKEN: VoteToken,{
    type Vote: VoteType<TYPES, COMMITMENT>;
    type VoteAccumulator: Accumulator2<TYPES, COMMITMENT, Self::Vote>;

    // Required methods
    fn from_signatures_and_commitment(
        signatures: AssembledSignature<TYPES>,
        vote: Self::Vote
    ) -> Self;
    fn view_number(&self) -> TIME;
    fn signatures(&self) -> AssembledSignature<TYPES>;
    fn leaf_commitment(&self) -> COMMITMENT;
    fn set_leaf_commitment(&mut self, commitment: COMMITMENT);
    fn is_genesis(&self) -> bool;
    fn genesis() -> Self;
}

A certificate of some property which has been signed by a quroum of nodes.

Required Associated Types

type Vote: VoteType<TYPES, COMMITMENT>

VoteType that is used in this certificate

type VoteAccumulator: Accumulator2<TYPES, COMMITMENT, Self::Vote>

Accumulator type to accumulate votes.

Required Methods

fn from_signatures_and_commitment( signatures: AssembledSignature<TYPES>, vote: Self::Vote ) -> Self

Build a QC from the threshold signature and commitment

fn view_number(&self) -> TIME

Get the view number.

fn signatures(&self) -> AssembledSignature<TYPES>

Get signatures.

fn leaf_commitment(&self) -> COMMITMENT

Get the leaf commitment.

fn set_leaf_commitment(&mut self, commitment: COMMITMENT)

Set the leaf commitment.

fn is_genesis(&self) -> bool

Get whether the certificate is for the genesis block.

fn genesis() -> Self

To be used only for generating the genesis quorum certificate; will fail if used anywhere else