Module hotshot_types::traits::metrics

The Metrics trait is used to collect information from multiple components in the entire system.

This trait can be used to spawn the following traits:

Counter: an ever-increasing value (example usage: total bytes send/received)

Gauge: a value that store the latest value, and can go up and down (example usage: amount of users logged in)

Histogram: stores multiple float values based for a graph (example usage: CPU %)

Label: Stores the last string (example usage: current version, network online/offline)

Structs

NoMetrics

Use this if you’re not planning to use any metrics. All methods are implemented as a no-op

Traits

Counter

An ever-incrementing counter

Gauge

A gauge that stores the latest value.

Histogram

A histogram which will record a series of points.

Label

A label that stores the last string value.

Metrics

The metrics type.