Trait hotshot_types::traits::network::CommunicationChannel

pub trait CommunicationChannel<TYPES: NodeType>: Clone + Debug + Send + Sync + 'static {
    type NETWORK;

    // Required methods
    fn wait_for_ready<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pause(&self);
    fn resume(&self);
    fn is_ready<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shut_down<'a, 'b>(&'a self) -> BoxSyncFuture<'b, ()>
       where Self: 'b,
             'a: 'b;
    fn broadcast_message<'life0, 'life1, 'async_trait>(
        &'life0 self,
        message: Message<TYPES>,
        election: &'life1 TYPES::Membership
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn direct_message<'life0, 'async_trait>(
        &'life0 self,
        message: Message<TYPES>,
        recipient: TYPES::SignatureKey
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn recv_msgs<'a, 'b>(
        &'a self,
        transmit_type: TransmitType
    ) -> BoxSyncFuture<'b, Result<Vec<Message<TYPES>>, NetworkError>>
       where Self: 'b,
             'a: 'b;

    // Provided methods
    fn queue_node_lookup<'life0, 'async_trait>(
        &'life0 self,
        _view_number: ViewNumber,
        _pk: TYPES::SignatureKey
    ) -> Pin<Box<dyn Future<Output = Result<(), UnboundedSendError<Option<(ViewNumber, TYPES::SignatureKey)>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn inject_consensus_info<'life0, 'async_trait>(
        &'life0 self,
        _event: ConsensusIntentEvent<TYPES::SignatureKey>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
}

API for interacting directly with a consensus committee intended to be implemented for both DA and for validating consensus committees

Required Associated Types

type NETWORK

Underlying Network implementation’s type

Required Methods

fn wait_for_ready<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Blocks until node is successfully initialized into the network

fn pause(&self)

Pauses the underlying network

fn resume(&self)

Resumes the underlying network

fn is_ready<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

checks if the network is ready nonblocking

fn shut_down<'a, 'b>(&'a self) -> BoxSyncFuture<'b, ()>where Self: 'b, 'a: 'b,

Shut down this network. Afterwards this network should no longer be used.

This should also cause other functions to immediately return with a NetworkError

fn broadcast_message<'life0, 'life1, 'async_trait>( &'life0 self, message: Message<TYPES>, election: &'life1 TYPES::Membership ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

broadcast message to those listening on the communication channel blocking

fn direct_message<'life0, 'async_trait>( &'life0 self, message: Message<TYPES>, recipient: TYPES::SignatureKey ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Sends a direct message to a specific node blocking

fn recv_msgs<'a, 'b>( &'a self, transmit_type: TransmitType ) -> BoxSyncFuture<'b, Result<Vec<Message<TYPES>>, NetworkError>>where Self: 'b, 'a: 'b,

Moves out the entire queue of received messages of ’transmit_type`

Will unwrap the underlying NetworkMessage blocking

Provided Methods

fn queue_node_lookup<'life0, 'async_trait>( &'life0 self, _view_number: ViewNumber, _pk: TYPES::SignatureKey ) -> Pin<Box<dyn Future<Output = Result<(), UnboundedSendError<Option<(ViewNumber, TYPES::SignatureKey)>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

queues looking up a node

fn inject_consensus_info<'life0, 'async_trait>( &'life0 self, _event: ConsensusIntentEvent<TYPES::SignatureKey> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Injects consensus data such as view number into the networking implementation blocking

Object Safety

This trait is not object safe.