Trait hotshot_types::traits::network::ConnectedNetwork

pub trait ConnectedNetwork<M: NetworkMsg, K: SignatureKey + 'static>: Clone + Send + Sync + 'static {

    // Required methods
    fn pause(&self);
    fn resume(&self);
    fn wait_for_ready<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn is_ready<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shut_down<'a, 'b>(&'a self) -> BoxSyncFuture<'b, ()>
       where Self: 'b,
             'a: 'b;
    fn broadcast_message<'life0, 'async_trait, VER>(
        &'life0 self,
        message: M,
        recipients: BTreeSet<K>,
        bind_version: VER
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where VER: 'async_trait + StaticVersionType + 'static,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn da_broadcast_message<'life0, 'async_trait, VER>(
        &'life0 self,
        message: M,
        recipients: BTreeSet<K>,
        bind_version: VER
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where VER: 'async_trait + StaticVersionType + 'static,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn direct_message<'life0, 'async_trait, VER>(
        &'life0 self,
        message: M,
        recipient: K,
        bind_version: VER
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where VER: 'async_trait + StaticVersionType + 'static,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn recv_msgs<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<M>, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided methods
    fn vid_broadcast_message<'life0, 'async_trait, VER>(
        &'life0 self,
        messages: HashMap<K, M>,
        bind_version: VER
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where VER: 'async_trait + StaticVersionType + 'static,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn request_data<'life0, 'async_trait, TYPES, VER>(
        &'life0 self,
        _request: M,
        _recipient: K,
        _bind_version: VER
    ) -> Pin<Box<dyn Future<Output = Result<ResponseMessage<TYPES>, NetworkError>> + Send + 'async_trait>>
       where TYPES: 'async_trait + NodeType,
             VER: 'async_trait + StaticVersionType + 'static,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn spawn_request_receiver_task<'life0, 'async_trait, VER>(
        &'life0 self,
        _bind_version: VER
    ) -> Pin<Box<dyn Future<Output = Option<Receiver<(M, ResponseChannel<M>)>>> + Send + 'async_trait>>
       where VER: 'async_trait + StaticVersionType + 'static,
             Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn queue_node_lookup<'life0, 'async_trait>(
        &'life0 self,
        _view_number: ViewNumber,
        _pk: K
    ) -> Pin<Box<dyn Future<Output = Result<(), UnboundedSendError<Option<(ViewNumber, K)>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn inject_consensus_info<'life0, 'async_trait>(
        &'life0 self,
        _event: ConsensusIntentEvent<K>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait { ... }
    fn update_view(&self, _view: u64) { ... }

}

represents a networking implmentration exposes low level API for interacting with a network intended to be implemented for libp2p, the centralized server, and memory network

Required Methods

fn pause(&self)

Pauses the underlying network

fn resume(&self)

Resumes the underlying network

fn wait_for_ready<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Blocks until the network is successfully initialized

fn is_ready<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = bool> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

checks if the network is ready nonblocking

fn shut_down<'a, 'b>(&'a self) -> BoxSyncFuture<'b, ()>
where Self: 'b, 'a: 'b,

Blocks until the network is shut down then returns true

fn broadcast_message<'life0, 'async_trait, VER>( &'life0 self, message: M, recipients: BTreeSet<K>, bind_version: VER ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
where VER: 'async_trait + StaticVersionType + 'static, Self: 'async_trait, 'life0: 'async_trait,

broadcast message to some subset of nodes blocking

fn da_broadcast_message<'life0, 'async_trait, VER>( &'life0 self, message: M, recipients: BTreeSet<K>, bind_version: VER ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
where VER: 'async_trait + StaticVersionType + 'static, Self: 'async_trait, 'life0: 'async_trait,

broadcast a message only to a DA committee blocking

fn direct_message<'life0, 'async_trait, VER>( &'life0 self, message: M, recipient: K, bind_version: VER ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
where VER: 'async_trait + StaticVersionType + 'static, Self: 'async_trait, 'life0: 'async_trait,

Sends a direct message to a specific node blocking

fn recv_msgs<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<Vec<M>, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Receive one or many messages from the underlying network.

Errors

If there is a network-related failure.

Provided Methods

fn vid_broadcast_message<'life0, 'async_trait, VER>( &'life0 self, messages: HashMap<K, M>, bind_version: VER ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
where VER: 'async_trait + StaticVersionType + 'static, Self: 'async_trait, 'life0: 'async_trait,

send messages with vid shares to its recipients blocking

fn request_data<'life0, 'async_trait, TYPES, VER>( &'life0 self, _request: M, _recipient: K, _bind_version: VER ) -> Pin<Box<dyn Future<Output = Result<ResponseMessage<TYPES>, NetworkError>> + Send + 'async_trait>>
where TYPES: 'async_trait + NodeType, VER: 'async_trait + StaticVersionType + 'static, Self: 'async_trait, 'life0: 'async_trait,

Ask request the network for some data. Returns the request ID for that data, The ID returned can be used for cancelling the request

fn spawn_request_receiver_task<'life0, 'async_trait, VER>( &'life0 self, _bind_version: VER ) -> Pin<Box<dyn Future<Output = Option<Receiver<(M, ResponseChannel<M>)>>> + Send + 'async_trait>>
where VER: 'async_trait + StaticVersionType + 'static, Self: 'async_trait, 'life0: 'async_trait,

Spawn a request task in the given network layer. If it supports Request and responses it will return the receiving end of a channel. Requests the network receives will be sent over this channel along with a return channel to send the response back to.

Returns `None`` if network does not support handling requests

fn queue_node_lookup<'life0, 'async_trait>( &'life0 self, _view_number: ViewNumber, _pk: K ) -> Pin<Box<dyn Future<Output = Result<(), UnboundedSendError<Option<(ViewNumber, K)>>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

queues lookup of a node

fn inject_consensus_info<'life0, 'async_trait>( &'life0 self, _event: ConsensusIntentEvent<K> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Injects consensus data such as view number into the networking implementation blocking Ideally we would pass in the Time type, but that requires making the entire trait generic over NodeType

fn update_view(&self, _view: u64)

handles view update

Object Safety

This trait is not object safe.