Module hotshot_types::traits::node_implementation

Composite trait for node behavior

This module defines the NodeImplementation trait, which is a composite trait used for describing the overall behavior of a node, as a composition of implementations of the node trait.

Structs

ChannelMaps

Channels for sending/recv-ing proposals and votes.

SendToTasks

metadata for sending information to the leader, replica, or DA committee member.

SequencingExchanges

Implementes SequencingExchangesType.

ValidatingExchanges

Implements ValidatingExchangesType.

ViewQueue

struct containing messages for a view to send to a replica or DA committee member.

Traits

ExchangesType

Contains the protocols for exchanging proposals and votes.

NodeImplementation

Node implementation aggregate trait

NodeType

Trait with all the type definitions that are used in the current hotshot setup.

SequencingExchangesType

An ExchangesType for sequencing consensus.

TestableNodeImplementation

extra functions required on a node implementation to be usable by hotshot-testing

ValidatingExchangesType

An ExchangesType for validating consensus.

Type Definitions

CommitteeCommChannel

Communication channel for CommitteeProposalType and [DAVote].

CommitteeEx

Alias for the [CommitteeExchange] type.

CommitteeMembership

Protocol for determining membership in a DA committee.

CommitteeNetwork

Type for the underlying committee ConnectedNetwork that will be shared (for now) b/t Communication Channels

CommitteeProposalType

A proposal to provide data availability for a new leaf.

CommitteeVote

A vote on a [ComitteeProposal].

ProcessedConsensusMessageType 🔒

Alias for the [ProcessedConsensusMessage] type of a NodeImplementation.

QuorumCommChannel

Communication channel for QuorumProposalType and [QuorumVote].

QuorumEx

Alias for the [QuorumExchange] type.

QuorumMembership

Protocol for determining membership in a consensus committee.

QuorumNetwork

Type for the underlying quorum ConnectedNetwork that will be shared (for now) b/t Communication Channels

QuorumProposalType

A proposal to append a new leaf to the log which is output by consensus.

QuorumVoteType

A vote on a QuorumProposalType.

SequencingQuorumEx

Alias for the [CommitteeExchange] type for sequencing consensus.

ValidatingQuorumEx

Alias for the [CommitteeExchange] type for validating consensus.

ViewSyncCommChannel

Communication channel for ViewSyncProposalType and [ViewSyncVote].

ViewSyncEx

Alias for the [ViewSyncExchange] type.

ViewSyncMembership

Protocol for determining membership in a view sync committee.

ViewSyncNetwork

Type for the underlying view sync ConnectedNetwork that will be shared (for now) b/t Communication Channels

ViewSyncProposalType

A proposal to sync the view.

ViewSyncVoteType

A vote on a [ViewSyncProposal].