Trait hotshot_types::traits::node_implementation::ExchangesType

pub trait ExchangesType<CONSENSUS: ConsensusType, TYPES: NodeType<ConsensusType = CONSENSUS>, LEAF: LeafType<NodeType = TYPES>, MESSAGE: NetworkMsg>: Send + Sync {
    type QuorumExchange: QuorumExchangeType<TYPES, LEAF, MESSAGE> + Clone + Debug;
    type Networks;

    // Required methods
    fn create(
        keys: Vec<TYPES::SignatureKey>,
        config: TYPES::ElectionConfigType,
        networks: Self::Networks,
        pk: TYPES::SignatureKey,
        sk: <TYPES::SignatureKey as SignatureKey>::PrivateKey,
        ek: KeyPair
    ) -> Self;
    fn quorum_exchange(&self) -> &Self::QuorumExchange;
    fn wait_for_networks_ready<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn shut_down_networks<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Contains the protocols for exchanging proposals and votes.

Required Associated Types

type QuorumExchange: QuorumExchangeType<TYPES, LEAF, MESSAGE> + Clone + Debug

Protocol for exchanging quorum proposals and votes.

type Networks

Networking implementations for quorum and committee exchanges, the latter of which is only applicable for sequencing consensus.

Required Methods

fn create( keys: Vec<TYPES::SignatureKey>, config: TYPES::ElectionConfigType, networks: Self::Networks, pk: TYPES::SignatureKey, sk: <TYPES::SignatureKey as SignatureKey>::PrivateKey, ek: KeyPair ) -> Self

Create all exchanges.

fn quorum_exchange(&self) -> &Self::QuorumExchange

Get the quorum exchange.

fn wait_for_networks_ready<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Block the underlying networking interfaces until node is successfully initialized into the networks.

fn shut_down_networks<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Shut down the the underlying networking interfaces.