Trait hotshot_types::traits::node_implementation::NodeType

pub trait NodeType: Clone + Copy + Debug + Hash + PartialEq + Eq + PartialOrd + Ord + Default + Serialize + for<'de> Deserialize<'de> + Send + Sync + 'static {
    type Time: ConsensusTime;
    type BlockHeader: BlockHeader<Self>;
    type BlockPayload: BlockPayload<Transaction = Self::Transaction>;
    type SignatureKey: SignatureKey;
    type Transaction: Transaction;
    type ElectionConfigType: ElectionConfig;
    type InstanceState: InstanceState;
    type ValidatedState: ValidatedState<Self, Instance = Self::InstanceState, Time = Self::Time>;
    type Membership: Membership<Self>;
}

Trait with all the type definitions that are used in the current hotshot setup.

Required Associated Types

type Time: ConsensusTime

The time type that this hotshot setup is using.

This should be the same Time that ValidatedState::Time is using.

type BlockHeader: BlockHeader<Self>

The block header type that this hotshot setup is using.

type BlockPayload: BlockPayload<Transaction = Self::Transaction>

The block type that this hotshot setup is using.

This should be the same block that ValidatedState::BlockPayload is using.

type SignatureKey: SignatureKey

The signature key that this hotshot setup is using.

type Transaction: Transaction

The transaction type that this hotshot setup is using.

This should be equal to BlockPayload::Transaction

type ElectionConfigType: ElectionConfig

The election config type that this hotshot setup is using.

type InstanceState: InstanceState

The instance-level state type that this hotshot setup is using.

type ValidatedState: ValidatedState<Self, Instance = Self::InstanceState, Time = Self::Time>

The validated state type that this hotshot setup is using.

type Membership: Membership<Self>

Membership used for this implementation

Object Safety

This trait is not object safe.