Trait hotshot_types::traits::node_implementation::TestableNodeImplementation

pub trait TestableNodeImplementation<TYPES: NodeType>: NodeImplementation<TYPES> {
    type CommitteeElectionConfig;

    // Required methods
    fn committee_election_config_generator(
    ) -> Box<dyn Fn(u64, u64) -> Self::CommitteeElectionConfig + 'static>;
    fn state_create_random_transaction(
        state: Option<&TYPES::ValidatedState>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <TYPES::BlockPayload as BlockPayload>::Transaction;
    fn leaf_create_random_transaction(
        leaf: &Leaf<TYPES>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <TYPES::BlockPayload as BlockPayload>::Transaction;
    fn block_genesis() -> TYPES::BlockPayload;
    fn txn_count(block: &TYPES::BlockPayload) -> u64;
    fn construct_tmp_storage() -> Result<Self::Storage, StorageError>;
    fn get_full_state<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = StorageState<TYPES>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn gen_networks(
        expected_node_count: usize,
        num_bootstrap: usize,
        da_committee_size: usize,
        reliability_config: Option<Box<dyn NetworkReliability>>,
        secondary_network_delay: Duration
    ) -> Box<dyn Fn(u64) -> (Arc<Self::QuorumNetwork>, Arc<Self::QuorumNetwork>)>;
}

extra functions required on a node implementation to be usable by hotshot-testing

Required Associated Types

type CommitteeElectionConfig

Election config for the DA committee

Required Methods

fn committee_election_config_generator( ) -> Box<dyn Fn(u64, u64) -> Self::CommitteeElectionConfig + 'static>

Generates a committee-specific election

fn state_create_random_transaction( state: Option<&TYPES::ValidatedState>, rng: &mut dyn RngCore, padding: u64 ) -> <TYPES::BlockPayload as BlockPayload>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn leaf_create_random_transaction( leaf: &Leaf<TYPES>, rng: &mut dyn RngCore, padding: u64 ) -> <TYPES::BlockPayload as BlockPayload>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn block_genesis() -> TYPES::BlockPayload

generate a genesis block

fn txn_count(block: &TYPES::BlockPayload) -> u64

the number of transactions in a block

fn construct_tmp_storage() -> Result<Self::Storage, StorageError>

Create ephemeral storage Will be deleted/lost immediately after storage is dropped

Errors

Errors if it is not possible to construct temporary storage.

fn get_full_state<'life0, 'async_trait>( storage: &'life0 Self::Storage ) -> Pin<Box<dyn Future<Output = StorageState<TYPES>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the full internal state. This is useful for debugging.

fn gen_networks( expected_node_count: usize, num_bootstrap: usize, da_committee_size: usize, reliability_config: Option<Box<dyn NetworkReliability>>, secondary_network_delay: Duration ) -> Box<dyn Fn(u64) -> (Arc<Self::QuorumNetwork>, Arc<Self::QuorumNetwork>)>

Generate the communication channels for testing

Object Safety

This trait is not object safe.