Trait hotshot_types::traits::qc::QuorumCertificate

pub trait QuorumCertificate<A: AggregateableSignatureSchemes + Serialize + for<'a> Deserialize<'a>> {
    type QCProverParams: Serialize + for<'a> Deserialize<'a>;
    type QCVerifierParams: Serialize + for<'a> Deserialize<'a>;
    type MessageLength: ArrayLength<A::MessageUnit>;
    type QC;
    type QuorumSize;

    // Required methods
    fn sign<R: CryptoRng + RngCore>(
        agg_sig_pp: &A::PublicParameter,
        message: &GenericArray<A::MessageUnit, Self::MessageLength>,
        sk: &A::SigningKey,
        prng: &mut R
    ) -> Result<A::Signature, PrimitivesError>;
    fn assemble(
        qc_pp: &Self::QCProverParams,
        signers: &BitSlice,
        sigs: &[A::Signature]
    ) -> Result<Self::QC, PrimitivesError>;
    fn check(
        qc_vp: &Self::QCVerifierParams,
        message: &GenericArray<A::MessageUnit, Self::MessageLength>,
        qc: &Self::QC
    ) -> Result<Self::QuorumSize, PrimitivesError>;
    fn trace(
        qc_vp: &Self::QCVerifierParams,
        message: &GenericArray<A::MessageUnit, Self::MessageLength>,
        qc: &Self::QC
    ) -> Result<Vec<A::VerificationKey>, PrimitivesError>;
}

Trait for validating a QC built from different signatures on the same message

Required Associated Types

type QCProverParams: Serialize + for<'a> Deserialize<'a>

Public parameters for generating the QC E.g: snark proving/verifying keys, list of (or pointer to) public keys stored in the smart contract.

type QCVerifierParams: Serialize + for<'a> Deserialize<'a>

Public parameters for validating the QC E.g: verifying keys, stake table commitment

type MessageLength: ArrayLength<A::MessageUnit>

Allows to fix the size of the message at compilation time.

type QC

Type of the actual quorum certificate object

type QuorumSize

Type of the quorum size (e.g. number of votes or accumulated weight of signatures)

Required Methods

fn sign<R: CryptoRng + RngCore>( agg_sig_pp: &A::PublicParameter, message: &GenericArray<A::MessageUnit, Self::MessageLength>, sk: &A::SigningKey, prng: &mut R ) -> Result<A::Signature, PrimitivesError>

Produces a partial signature on a message with a single user signing key NOTE: the original message (vote) should be prefixed with the hash of the stake table.

agg_sig_pp - public parameters for aggregate signature

message - message to be signed

sk - user signing key

returns - a “simple” signature

Errors

Should return error if the underlying signature scheme fail to sign.

fn assemble( qc_pp: &Self::QCProverParams, signers: &BitSlice, sigs: &[A::Signature] ) -> Result<Self::QC, PrimitivesError>

Computes an aggregated signature from a set of partial signatures and the verification keys involved

qc_pp - public parameters for generating the QC

signers - a bool vector indicating the list of verification keys corresponding to the set of partial signatures

sigs - partial signatures on the same message

Errors

Will return error if some of the partial signatures provided are invalid or the number of partial signatures / verifications keys are different.

fn check( qc_vp: &Self::QCVerifierParams, message: &GenericArray<A::MessageUnit, Self::MessageLength>, qc: &Self::QC ) -> Result<Self::QuorumSize, PrimitivesError>

Checks an aggregated signature over some message provided as input

qc_vp - public parameters for validating the QC

message - message to check the aggregated signature against

qc - quroum certificate

returns - the quorum size if the qc is valid, an error otherwise.

Errors

Return error if the QC is invalid, either because accumulated weight didn’t exceed threshold, or some partial signatures are invalid.

fn trace( qc_vp: &Self::QCVerifierParams, message: &GenericArray<A::MessageUnit, Self::MessageLength>, qc: &Self::QC ) -> Result<Vec<A::VerificationKey>, PrimitivesError>

Trace the list of signers given a qc.

Errors

Return error if the inputs mismatch (e.g. wrong verifier parameter or original message).