hotshot_types::traits
Module signature_key

Minimal compatibility over public key signatures

Traits

BuilderSignatureKey

Builder Signature Key trait with minimal requirements

PrivateSignatureKey

Trait for abstracting private signature key

SignatureKey

Trait for abstracting public key signatures Self is the public key type

StakeTableEntryType

Type representing stake table entries in a StakeTable

Functions

aggregate_block_info_data 🔒

Aggregate all inputs used for signature over block info

aggregate_fee_data 🔒

Aggregate all inputs used for signature over fee data

aggregate_fee_data_marketplace 🔒

Aggregate all inputs used for signature over fee data