Trait hotshot_types::traits::signature_key::BuilderSignatureKey

pub trait BuilderSignatureKey: Send + Sync + Clone + Sized + Debug + Hash + Serialize + for<'a> Deserialize<'a> + PartialEq + Eq + PartialOrd + Ord + Display {
    type BuilderPrivateKey: Send + Sync + Sized + Clone + Debug + Eq + Serialize + for<'a> Deserialize<'a> + Hash;
    type BuilderSignature: Send + Sync + Sized + Clone + Debug + Eq + Serialize + for<'a> Deserialize<'a> + Hash;
    type SignError: Error + Send + Sync;

    // Required methods
    fn validate_builder_signature(
        &self,
        signature: &Self::BuilderSignature,
        data: &[u8]
    ) -> bool;
    fn sign_builder_message(
        private_key: &Self::BuilderPrivateKey,
        data: &[u8]
    ) -> Result<Self::BuilderSignature, Self::SignError>;
    fn generated_from_seed_indexed(
        seed: [u8; 32],
        index: u64
    ) -> (Self, Self::BuilderPrivateKey);
}

Builder Signature Key trait with minimal requirements

Required Associated Types

type BuilderPrivateKey: Send + Sync + Sized + Clone + Debug + Eq + Serialize + for<'a> Deserialize<'a> + Hash

The type of the keys builder would use to sign its messages

type BuilderSignature: Send + Sync + Sized + Clone + Debug + Eq + Serialize + for<'a> Deserialize<'a> + Hash

The type of the signature builder would use to sign its messages

type SignError: Error + Send + Sync

Type of error that can occur when signing data

Required Methods

fn validate_builder_signature( &self, signature: &Self::BuilderSignature, data: &[u8] ) -> bool

validate the message with the builder’s public key

fn sign_builder_message( private_key: &Self::BuilderPrivateKey, data: &[u8] ) -> Result<Self::BuilderSignature, Self::SignError>

sign the message with the builder’s private key

Errors

If unable to sign the data with the key

fn generated_from_seed_indexed( seed: [u8; 32], index: u64 ) -> (Self, Self::BuilderPrivateKey)

Generate a new key pair

Object Safety

This trait is not object safe.