Trait hotshot_types::traits::state::TestableState

pub trait TestableState: Statewhere
    <Self as State>::BlockType: TestableBlock,{
    // Required method
    fn create_random_transaction(
        state: Option<&Self>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <Self::BlockType as Block>::Transaction;
}

extra functions required on state to be usable by hotshot-testing

Required Methods

fn create_random_transaction( state: Option<&Self>, rng: &mut dyn RngCore, padding: u64 ) -> <Self::BlockType as Block>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction