hotshot_types
Module vid

This module provides:

an opaque constructor [vid_scheme] that returns a new instance of a VID scheme.

type aliases VidCommitment, VidCommon, VidShare for [VidScheme] assoc types.

Purpose: the specific choice of VID scheme is an implementation detail. This crate and all downstream crates should talk to the VID scheme only via the traits exposed here.

Structs

KZG_SRS 🔒

SRS comment

LargeRangeProofType

Newtype wrapper for a large payload range proof.

MEMOIZED_MAPPING_ADVZ_SCHEME 🔒

SmallRangeProofType

Newtype wrapper for a small payload range proof.

VidSchemeType

Newtype wrapper for a VID scheme type that impls [VidScheme], [PayloadProver], [Precomputable].

Functions

advz_scheme

memoized_flush_advz_scheme

memoized_original_advz_scheme

VID scheme constructor.

stmt_conversion 🔒

Convert a [Statement<'_, VidSchemeType>] to a [Statement<'_, Advz>].

vid_disperse_conversion 🔒

Convert a [VidDisperse<Advz>] to a [VidDisperse<VidSchemeType>].

Type Aliases

Advz 🔒

Internal Jellyfish VID scheme

E 🔒

Private type alias for the EC pairing type parameter for Advz.

H 🔒

Private type alias for the hash type parameter for Advz.

VidCommitment

VID commitment type

VidCommon

VID common type

VidProposal

VID proposal type

VidShare

VID share type