hotshot_types::vid::advz
Function memoized_original_advz_scheme

pub fn memoized_original_advz_scheme(num_storage_nodes: usize) -> ADVZScheme

VID scheme constructor.

Returns an opaque type that impls jellyfish traits: [VidScheme], [PayloadProver], [Precomputable].

Rust forbids naming impl Trait in return types

Due to Rust limitations the return type of [vid_scheme] is a newtype wrapper [VidSchemeType] that impls the above traits.

We prefer that the return type of [vid_scheme] be impl Trait for the above traits. But the ability to name an impl Trait return type is currently missing from Rust:

Naming impl trait in return types - Impl trait initiative

RFC: Type alias impl trait (TAIT)

Panics

When the construction fails for the underlying VID scheme.