Trait hotshot_types::vote::Accumulator2

pub trait Accumulator2<TYPES: NodeType, COMMITTABLE: Committable + Serialize + Clone, VOTE: VoteType<TYPES, Commitment<COMMITTABLE>>>: Sized {
    // Required method
    fn append(
        self,
        vote: VOTE,
        vote_node_id: usize,
        stake_table_entries: Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>
    ) -> Either<Self, AssembledSignature<TYPES>>;
}

Accumulator trait used to accumulate votes into an AssembledSignature

Required Methods

fn append( self, vote: VOTE, vote_node_id: usize, stake_table_entries: Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry> ) -> Either<Self, AssembledSignature<TYPES>>

Append 1 vote to the accumulator. If the threshold is not reached, return the accumulator, else return the AssembledSignature Only called from inside accumulate_internal