Trait hotshot_types::vote::Certificate

pub trait Certificate<TYPES: NodeType>: HasViewNumber<TYPES> {
    type Voteable: Voteable;
    type Threshold: Threshold<TYPES>;

    // Required methods
    fn create_signed_certificate(
        vote_commitment: Commitment<Self::Voteable>,
        data: Self::Voteable,
        sig: <TYPES::SignatureKey as SignatureKey>::QCType,
        view: TYPES::Time
    ) -> Self;
    fn is_valid_cert<MEMBERSHIP: Membership<TYPES>>(
        &self,
        membership: &MEMBERSHIP
    ) -> bool;
    fn threshold<MEMBERSHIP: Membership<TYPES>>(membership: &MEMBERSHIP) -> u64;
    fn get_data(&self) -> &Self::Voteable;
    fn get_data_commitment(&self) -> Commitment<Self::Voteable>;
}

The certificate formed from the collection of signatures a committee. The committee is defined by the Membership associated type. The votes all must be over the Commitment associated type.

Required Associated Types

type Voteable: Voteable

The data commitment this certificate certifies.

type Threshold: Threshold<TYPES>

Threshold Functions

Required Methods

fn create_signed_certificate( vote_commitment: Commitment<Self::Voteable>, data: Self::Voteable, sig: <TYPES::SignatureKey as SignatureKey>::QCType, view: TYPES::Time ) -> Self

Build a certificate from the data commitment and the quorum of signers

fn is_valid_cert<MEMBERSHIP: Membership<TYPES>>( &self, membership: &MEMBERSHIP ) -> bool

Checks if the cert is valid

fn threshold<MEMBERSHIP: Membership<TYPES>>(membership: &MEMBERSHIP) -> u64

Returns the amount of stake needed to create this certificate

fn get_data(&self) -> &Self::Voteable

Get the commitment which was voted on

fn get_data_commitment(&self) -> Commitment<Self::Voteable>

Get the vote commitment which the votes commit to

Object Safety

This trait is not object safe.