Module hotshot_types::consensus

Provides the core consensus types

Re-exports

pub use crate::utils::View;

pub use crate::utils::ViewInner;

Structs

CommitmentAndMetadata

Alias for the block payload commitment and the associated metadata. The primary data needed in order to submit a proposal.

Consensus

A reference to the consensus algorithm

ConsensusMetricsValue

Contains several ConsensusMetrics that we’re interested in from the consensus interfaces

ConsensusReadLockGuard

A thin wrapper around RwLockReadGuard for Consensus that leaves debug traces when the lock is freed

ConsensusUpgradableReadLockGuard

A thin wrapper around RwLockUpgradableReadGuard for Consensus that leaves debug traces when the lock is freed or upgraded

ConsensusWriteLockGuard

A thin wrapper around RwLockWriteGuard for Consensus that leaves debug traces when the lock is freed

HotShotActionViews 🔒

A bundle of views that we have most recently performed some action

OuterConsensus

A thin wrapper around LockedConsensusState that helps debugging locks

Type Aliases

CommitmentMap

A type alias for HashMap<Commitment<T>, T>

LockedConsensusState

Type alias for consensus state wrapped in a lock.

VidShares

A type alias for BTreeMap<T::Time, HashMap<T::SignatureKey, Proposal<T, VidDisperseShare<T>>>>