hotshot_types

Module constants

configurable constants for hotshot

Constants

AUCTION_RESULTS_FETCH_TIMEOUT

timeout for fetching auction results from the solver

BUNDLE_FETCH_TIMEOUT

timeout for fetching bundles from builders

COMBINED_NETWORK_CACHE_SIZE

the number of messages to cache in the combined network

COMBINED_NETWORK_DELAY_DURATION

the default delay duration value in milliseconds of sending on the secondary in the combined networks

COMBINED_NETWORK_MIN_PRIMARY_FAILURES

the number of messages to attempt to send over the primary network before switching to prefer the secondary network

COMBINED_NETWORK_PRIMARY_CHECK_INTERVAL

the number of messages to send over the secondary network without delay before re-attempting the (presumed down) primary network

DEFAULT_UPGRADE_CONSTANTS

Default values for the upgrade constants

EVENT_CHANNEL_SIZE

Default channel size for consensus event sharing

EXTERNAL_EVENT_CHANNEL_SIZE

Default channel size for HotShot -> application communication

KAD_DEFAULT_REPUB_INTERVAL_SEC

the default kademlia record republication interval (in seconds)

LEGACY_BUILDER_MODULE

The tide module name for the legacy builder

LOOK_AHEAD

the number of views to gather information for ahead of time

MARKETPLACE_BUILDER_MODULE

The tide module name for the marketplace builder

ORCHESTRATOR_DEFAULT_NUM_ROUNDS

default number of rounds to run

ORCHESTRATOR_DEFAULT_TRANSACTIONS_PER_ROUND

default number of transactions per round

ORCHESTRATOR_DEFAULT_TRANSACTION_SIZE

default size of transactions

REQUEST_DATA_DELAY

The default network data request delay in milliseconds

SRS_DEGREE

For STAKE_TABLE_CAPACITY=200, the light client prover (a.k.a. hotshot-state-prover) would need to generate proof for a circuit of slightly below 2^20 gates. Thus we need to support this upperbounded degree in our Structured Reference String (SRS), the +2 is just an artifact from the jellyfish’s Plonk proof system.

TEST_UPGRADE_CONSTANTS

Default values for the upgrade constants to be used in testing