Module hotshot_types::data

Provides types useful for representing HotShot’s data structures

This module provides types for representing consensus internal state, such as leaves, HotShot’s version of a block, and proposals, messages upon which to reach the consensus.

Modules

null_block

Structs

DAProposal

A proposal to start providing data availability for a block.

InconsistentPayloadCommitmentSnafu 🔒

SNAFU context selector for the BlockError::InconsistentPayloadCommitment variant

InvalidBlockHeaderSnafu 🔒

SNAFU context selector for the BlockError::InvalidBlockHeader variant

InvalidTransactionLengthSnafu 🔒

SNAFU context selector for the BlockError::InvalidTransactionLength variant

Leaf

This is the consensus-internal analogous concept to a block, and it contains the block proper, as well as the hash of its parent Leaf. NOTE: State is constrained to implementing BlockContents, is TypeMap::BlockPayload

QuorumProposal

Proposal to append a block.

UpgradeProposal

A proposal to upgrade the network

VidDisperse

VID dispersal data

VidDisperseShare

VID share and associated metadata for a single node

ViewNumber

Type-safe wrapper around u64 so we know the thing we’re talking about is a view number.

Enums

BlockError

The error type for block and its transactions.

ViewChangeEvidence

Helper type to encapsulate the various ways that proposal certificates can be captured and stored.

Constants

_DERIVE_core_fmt_Debug_FOR_DAProposal 🔒

_DERIVE_core_fmt_Debug_FOR_QuorumProposal 🔒

_DERIVE_core_fmt_Debug_FOR_UpgradeProposal 🔒

_DERIVE_core_fmt_Debug_FOR_ViewChangeEvidence 🔒

Traits

TestableLeaf

Additional functions required to use a Leaf with hotshot-testing.

Functions

fake_commitment

Fake the thing a genesis block points to. Needed to avoid infinite recursion

random_commitment

create a random commitment

serialize_signature2

Serialization for the QC assembled signature