Trait hotshot_types::data::DeltasType

pub trait DeltasType<Block: Committable>: Clone + Debug + for<'a> Deserialize<'a> + PartialEq + Send + Serialize + Sync {
    type Error: Error;

    fn block_commitment(&self) -> Commitment<Block>;
    fn try_resolve(self) -> Result<Block, Self>;
    fn fill(&mut self, block: Block) -> Result<(), Self::Error>;
}

A state change encoded in a leaf.

DeltasType represents a block, but it may not contain the block in full. It is guaranteed to contain, at least, a cryptographic commitment to the block, and it provides an interface for resolving the commitment to a full block if the full block is available.

Required Associated Types

type Error: Error

Errors reported by this type.

Required Methods

fn block_commitment(&self) -> Commitment<Block>

Get a cryptographic commitment to the block represented by this delta.

fn try_resolve(self) -> Result<Block, Self>

Get the full block if it is available, otherwise return this object unchanged.

Errors

Returns the original DeltasType, unchanged, in an Err variant in the case where the full block is not currently available.

fn fill(&mut self, block: Block) -> Result<(), Self::Error>

Fill this DeltasType by providing a complete block.

After this function succeeds, try_resolve is guaranteed to return Ok(block).

Errors

Fails if block does not match self.block_commitment(), or if the block is not able to be stored for some implementation-defined reason.