hotshot_types::data
Trait TestableLeaf

pub trait TestableLeaf {
    type NodeType: NodeType;

    // Required method
    fn create_random_transaction(
        &self,
        rng: &mut dyn RngCore,
        padding: u64,
    ) -> <<Self::NodeType as NodeType>::BlockPayload as BlockPayload<Self::NodeType>>::Transaction;
}

Additional functions required to use a Leaf with hotshot-testing.

Required Associated Types

type NodeType: NodeType

Type of nodes participating in the network.

Required Methods

fn create_random_transaction( &self, rng: &mut dyn RngCore, padding: u64, ) -> <<Self::NodeType as NodeType>::BlockPayload as BlockPayload<Self::NodeType>>::Transaction

Create a transaction that can be added to the block contained in this leaf.