hotshot_types
Module event

Events that a HotShot instance can emit

Modules

error_adaptor

Utilities for converting between HotShotError and a string.

Structs

Event

A status event emitted by a HotShot instance

LeafInfo

Decided leaf with the corresponding state and VID info.

Enums

EventType

The type and contents of a status event emitted by a HotShot instance

HotShotAction

A list of actions that we track for nodes

Type Aliases

LeafChain

The chain of decided leaves with its corresponding state and VID info.