hotshot_types

Module light_client

Types and structs associated with light client state

Structs

GenericLightClientState

A light client state

GenericStakeTableState

Stake table state

LCV1StateSignatureRequestBody

The request body for light client V1 to send to the state relay server

LCV1StateSignaturesBundle

The state signatures bundle is a light client V1 state and its signatures collected

LCV2StateSignatureRequestBody

The request body for light client V2 to send to the state relay server

LCV2StateSignaturesBundle

The state signatures bundle is a light client V2 state and its signatures collected

LCV3StateSignatureRequestBody

The request body for light client V3 to send to the state relay server

LCV3StateSignaturesBundle

The state signatures bundle is a light client V3 state and its signatures collected

StateKeyPair

Key pairs for signing/verifying a light client state

Constants

DEFAULT_STAKE_TABLE_CAPACITY

Capacity of the stake table, used for light client

Traits

ToFieldsLightClientCompat

This trait is for light client use. It converts the stake table items into field elements. These items will then be digested into a part of the light client state.

Functions

hash_bytes_to_field

Type Aliases

CircuitField

Base field in the prover circuit

GenericLightClientStateMsg

LightClientState

Concrete type for light client state

LightClientStateMsg

Concreate type for light client state message to sign

StakeTableState

Concrete type for stake table state

StateSignKey

Signing key for signing a light client state

StateSignature

Signatures

StateSignatureScheme

Signature scheme

StateVerKey

Verification key for verifying state signatures