Module hotshot_types::message

Network message types

This module contains types used to represent the various types of messages that HotShot nodes can send among themselves.

Structs

Message

Incoming message

Messages

A wrapper type for implementing PassType on a vector of Message.

Proposal

Prepare qc from the leader

UpgradeLock

A lock for an upgrade certificate decided by HotShot, which doubles as PhantomData for an instance of the Versions trait.

Enums

DaConsensusMessage

Messages related to the sequencing consensus protocol for the DA committee.

DataMessage

TODO: Put DataResponse content in a Box to make enum smaller Messages related to sending data between nodes

GeneralConsensusMessage

Messages related to both validating and sequencing consensus.

MessageKind

Enum representation of any message type

MessagePurpose

A message type agnostic description of a message’s purpose

RecipientList

List of keys to send a message to, or broadcast to all known keys

SequencingMessage

Messages for sequencing consensus.

Functions

convert_proposal

Convert a Proposal by converting the underlying proposal type