hotshot_types

Module simple_certificate

Implementations of the simple certificate type. Used for Quorum, DA, and Timeout Certificates

Structs

LightClientStateUpdateCertificate

Type for light client state update certificate

OneHonestThreshold

Defines a threshold which is f + 1 (i.e at least one of the stake is honest)

SimpleCertificate

A certificate which can be created by aggregating many simple votes on the commitment.

SuccessThreshold

Defines a threshold which is 2f + 1 (Amount needed for Quorum)

UpgradeThreshold

Defines a threshold which is 0.9n + 1 (i.e. over 90% of the nodes with stake)

Traits

Threshold

Trait which allows use to inject different threshold calculations into a Certificate type

Type Aliases

DaCertificate

Type alias for a DaCertificate, which is a SimpleCertificate over DaData

DaCertificate2

Type alias for a DaCertificate2, which is a SimpleCertificate over DaData2

NextEpochQuorumCertificate2

Type alias for a QuorumCertificate2, which is a SimpleCertificate over QuorumData2

QuorumCertificate

Type alias for a QuorumCertificate, which is a SimpleCertificate over QuorumData

QuorumCertificate2

Type alias for a QuorumCertificate2, which is a SimpleCertificate over QuorumData2

TimeoutCertificate

Type alias for a Timeout certificate over a view number

TimeoutCertificate2

Type alias for a TimeoutCertificate2, which is a SimpleCertificate over TimeoutData2

UpgradeCertificate

Type alias for a UpgradeCertificate, which is a SimpleCertificate of UpgradeProposalData

ViewSyncCommitCertificate

Type alias for a ViewSyncCommit certificate over a view number

ViewSyncCommitCertificate2

Type alias for a ViewSyncCommitCertificate2, which is a SimpleCertificate over ViewSyncCommitData2

ViewSyncFinalizeCertificate

Type alias for a ViewSyncFinalize certificate over a view number

ViewSyncFinalizeCertificate2

Type alias for a ViewSyncFinalizeCertificate2, which is a SimpleCertificate over ViewSyncFinalizeData2

ViewSyncPreCommitCertificate

Type alias for a ViewSyncPreCommit certificate over a view number

ViewSyncPreCommitCertificate2

Type alias for a ViewSyncPreCommitCertificate2, which is a SimpleCertificate over ViewSyncPreCommitData2