Type Alias hotshot_types::simple_certificate::ViewSyncPreCommitCertificate2

pub type ViewSyncPreCommitCertificate2<TYPES> = SimpleCertificate<TYPES, ViewSyncPreCommitData2<TYPES>, OneHonestThreshold>;

Type alias for a ViewSyncPreCommitCertificate2, which is a SimpleCertificate over ViewSyncPreCommitData2

Aliased Type

struct ViewSyncPreCommitCertificate2<TYPES> {
    pub data: ViewSyncPreCommitData2<TYPES>,
    vote_commitment: Commitment<ViewSyncPreCommitData2<TYPES>>,
    pub view_number: <TYPES as NodeType>::View,
    pub signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QcType>,
    pub _pd: PhantomData<(TYPES, OneHonestThreshold)>,
}

Fields
data: ViewSyncPreCommitData2<TYPES>

The data this certificate is for. I.e the thing that was voted on to create this Certificate

vote_commitment: Commitment<ViewSyncPreCommitData2<TYPES>>

commitment of all the votes this cert should be signed over

view_number: <TYPES as NodeType>::View

Which view this QC relates to

signatures: Option<<<TYPES as NodeType>::SignatureKey as SignatureKey>::QcType>

assembled signature for certificate aggregation

_pd: PhantomData<(TYPES, OneHonestThreshold)>

phantom data for THRESHOLD and TYPES

Implementations

impl<TYPES: NodeType> ViewSyncPreCommitCertificate2<TYPES>

pub fn to_vsc(self) -> ViewSyncPreCommitCertificate<TYPES>

Convert a DaCertificate into a DaCertificate2