Module hotshot_types::traits

Common traits for the HotShot protocol

Re-exports

pub use block_contents::BlockPayload;

pub use block_contents::EncodeBytes;

pub use states::ValidatedState;

Modules

auction_results_provider

This module defines the interaction layer with the Solver via the AuctionResultsProvider trait, which handles connecting to, and fetching the allocation results from, the Solver.

block_contents

Abstraction over the contents of a block

consensus_api

Contains the ConsensusApi trait.

election

The election trait, used to decide which node is the leader and determine if a vote is valid.

metrics

The Metrics trait is used to collect information from multiple components in the entire system.

network

Network access compatibility

node_implementation

Composite trait for node behavior

qc

The quorum certificate (QC) trait is a certificate of a sufficient quorum of distinct parties voted for a message or statement.

signature_key

Minimal compatibility over public key signatures

stake_table

Trait for stake table data structures

states

Abstractions over the immutable instance-level state and the global state that blocks modify.

storage

Abstract storage type for storing DA proposals and VID shares