Module hotshot_types::traits::block_contents

Abstraction over the contents of a block

This module provides the Transaction, BlockPayload, and BlockHeader traits, which describe the behaviors that a block is expected to have.

Structs

BuilderFee

Information about builder fee for proposed block

Constants

GENESIS_VID_NUM_STORAGE_NODES

The number of storage nodes to use when computing the genesis VID commitment.

Traits

BlockHeader

Header of a block, which commits to a BlockPayload.

BlockPayload

Abstraction over the full contents of a block

EncodeBytes

Trait for structures that need to be unambiguously encoded as bytes.

TestableBlock

extra functions required on block to be usable by hotshot-testing

Transaction

Abstraction over any type of transaction. Used by BlockPayload.

Functions

vid_commitment

Compute the VID payload commitment. TODO(Gus) delete this function? TODO(Chengyu): use the version information