Trait hotshot_types::traits::block_contents::Block

pub trait Block: Serialize + Clone + Debug + Hash + PartialEq + Eq + Send + Sync + Committable + DeserializeOwned {
    type Error: Error + Debug + Send + Sync;
    type Transaction: Transaction;

    fn new() -> Self;
    fn add_transaction_raw(
        &self,
        tx: &Self::Transaction
    ) -> Result<Self, Self::Error>;
    fn contained_transactions(&self) -> HashSet<Commitment<Self::Transaction>>;
}

Abstraction over the full contents of a block

This trait encapsulates the behaviors that a block must have in order to be used by consensus:

Must have a predefined error type (Block::Error)

Must have a transaction type that can be compared for equality, serialized and serialized, sent between threads, and can have a hash produced of it

Must be able to be produced incrementally by appending transactions (add_transaction_raw)

Must be hashable

Required Associated Types

type Error: Error + Debug + Send + Sync

The error type for this type of block

type Transaction: Transaction

The type of the transitions we are applying

Required Methods

fn new() -> Self

Construct an empty or genesis block.

fn add_transaction_raw(
&self,
tx: &Self::Transaction
) -> Result<Self, Self::Error>

Attempts to add a transaction, returning an Error if it would result in a structurally invalid block

Errors

Should return an error if this transaction leads to an invalid block

fn contained_transactions(&self) -> HashSet<Commitment<Self::Transaction>>

returns hashes of all the transactions in this block TODO make this ordered with a vec