Trait hotshot_types::traits::block_contents::BlockHeader

pub trait BlockHeader: Serialize + Clone + Debug + Hash + PartialEq + Eq + Send + Sync + DeserializeOwned + Committable {
    type Payload: BlockPayload;

    // Required methods
    fn new(
        payload_commitment: VidCommitment,
        metadata: <Self::Payload as BlockPayload>::Metadata,
        parent_header: &Self
    ) -> Self;
    fn genesis(
    ) -> (Self, Self::Payload, <Self::Payload as BlockPayload>::Metadata);
    fn block_number(&self) -> u64;
    fn payload_commitment(&self) -> VidCommitment;
    fn metadata(&self) -> <Self::Payload as BlockPayload>::Metadata;
}

Header of a block, which commits to a BlockPayload.

Required Associated Types

type Payload: BlockPayload

Block payload associated with the commitment.

Required Methods

fn new( payload_commitment: VidCommitment, metadata: <Self::Payload as BlockPayload>::Metadata, parent_header: &Self ) -> Self

Build a header with the payload commitment, metadata, and parent header.

fn genesis() -> (Self, Self::Payload, <Self::Payload as BlockPayload>::Metadata)

Build the genesis header, payload, and metadata.

fn block_number(&self) -> u64

Get the block number.

fn payload_commitment(&self) -> VidCommitment

Get the payload commitment.

fn metadata(&self) -> <Self::Payload as BlockPayload>::Metadata

Get the metadata.