Trait hotshot_types::traits::election::CommitteeExchangeType

pub trait CommitteeExchangeType<TYPES: NodeType<ConsensusType = SequencingConsensus>, M: NetworkMsg>: ConsensusExchange<TYPES, M> {
    // Required methods
    fn sign_da_proposal(
        &self,
        block_commitment: &Commitment<TYPES::BlockType>
    ) -> EncodedSignature;
    fn sign_da_vote(
        &self,
        block_commitment: Commitment<TYPES::BlockType>
    ) -> (EncodedPublicKey, EncodedSignature);
    fn create_da_message<I: NodeImplementation<TYPES, ConsensusMessage = SequencingMessage<TYPES, I>>>(
        &self,
        justify_qc_commitment: Commitment<QuorumCertificate<TYPES, I::Leaf>>,
        block_commitment: Commitment<TYPES::BlockType>,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> CommitteeConsensusMessage<TYPES, I>
       where I::Exchanges: SequencingExchangesType<TYPES, Message<TYPES, I>>;
}

A ConsensusExchange where participants vote to provide availability for blobs of data.

Required Methods

fn sign_da_proposal( &self, block_commitment: &Commitment<TYPES::BlockType> ) -> EncodedSignature

Sign a DA proposal.

fn sign_da_vote( &self, block_commitment: Commitment<TYPES::BlockType> ) -> (EncodedPublicKey, EncodedSignature)

Sign a vote on DA proposal.

The block commitment and the type of the vote (DA) are signed, which is the minimum amount of information necessary for checking that this node voted on that block.

fn create_da_message<I: NodeImplementation<TYPES, ConsensusMessage = SequencingMessage<TYPES, I>>>( &self, justify_qc_commitment: Commitment<QuorumCertificate<TYPES, I::Leaf>>, block_commitment: Commitment<TYPES::BlockType>, current_view: TYPES::Time, vote_token: TYPES::VoteTokenType ) -> CommitteeConsensusMessage<TYPES, I>where I::Exchanges: SequencingExchangesType<TYPES, Message<TYPES, I>>,

Create a message with a vote on DA proposal.