Trait hotshot_types::traits::election::Membership

pub trait Membership<TYPES: NodeType>: Clone + Debug + Eq + PartialEq + Send + Sync + Hash + 'static {
    // Required methods
    fn new(
        eligible_leaders: Vec<PeerConfig<TYPES::SignatureKey>>,
        committee_members: Vec<PeerConfig<TYPES::SignatureKey>>,
        committee_topic: Topic
    ) -> Self;
    fn stake_table(
        &self,
        epoch: TYPES::Epoch
    ) -> Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>;
    fn committee_members(
        &self,
        view_number: TYPES::View,
        epoch: TYPES::Epoch
    ) -> BTreeSet<TYPES::SignatureKey>;
    fn committee_leaders(
        &self,
        view_number: TYPES::View,
        epoch: TYPES::Epoch
    ) -> BTreeSet<TYPES::SignatureKey>;
    fn stake(
        &self,
        pub_key: &TYPES::SignatureKey,
        epoch: TYPES::Epoch
    ) -> Option<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>;
    fn has_stake(
        &self,
        pub_key: &TYPES::SignatureKey,
        epoch: TYPES::Epoch
    ) -> bool;
    fn leader(
        &self,
        view_number: TYPES::View,
        epoch: TYPES::Epoch
    ) -> TYPES::SignatureKey;
    fn committee_topic(&self) -> Topic;
    fn total_nodes(&self, epoch: TYPES::Epoch) -> usize;
    fn success_threshold(&self) -> NonZeroU64;
    fn failure_threshold(&self) -> NonZeroU64;
    fn upgrade_threshold(&self) -> NonZeroU64;
}

A protocol for determining membership in and participating in a committee.

Required Methods

fn new( eligible_leaders: Vec<PeerConfig<TYPES::SignatureKey>>, committee_members: Vec<PeerConfig<TYPES::SignatureKey>>, committee_topic: Topic ) -> Self

Create a committee

fn stake_table( &self, epoch: TYPES::Epoch ) -> Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>

Get all participants in the committee (including their stake) for a specific epoch

fn committee_members( &self, view_number: TYPES::View, epoch: TYPES::Epoch ) -> BTreeSet<TYPES::SignatureKey>

Get all participants in the committee for a specific view for a specific epoch

fn committee_leaders( &self, view_number: TYPES::View, epoch: TYPES::Epoch ) -> BTreeSet<TYPES::SignatureKey>

Get all leaders in the committee for a specific view for a specific epoch

fn stake( &self, pub_key: &TYPES::SignatureKey, epoch: TYPES::Epoch ) -> Option<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>

Get the stake table entry for a public key, returns None if the key is not in the table for a specific epoch

fn has_stake(&self, pub_key: &TYPES::SignatureKey, epoch: TYPES::Epoch) -> bool

See if a node has stake in the committee in a specific epoch

fn leader( &self, view_number: TYPES::View, epoch: TYPES::Epoch ) -> TYPES::SignatureKey

The leader of the committee for view view_number in an epoch epoch.

fn committee_topic(&self) -> Topic

Get the network topic for the committee

fn total_nodes(&self, epoch: TYPES::Epoch) -> usize

Returns the number of total nodes in the committee in an epoch epoch

fn success_threshold(&self) -> NonZeroU64

Returns the threshold for a specific Membership implementation

fn failure_threshold(&self) -> NonZeroU64

Returns the threshold for a specific Membership implementation

fn upgrade_threshold(&self) -> NonZeroU64

Returns the threshold required to upgrade the network protocol

Object Safety

This trait is not object safe.