Trait hotshot_types::traits::election::Membership

pub trait Membership<TYPES: NodeType>: Clone + Debug + Eq + PartialEq + Send + Sync + Hash + 'static {

    // Required methods
    fn default_election_config(
        num_nodes_with_stake: u64,
        num_nodes_without_stake: u64
    ) -> TYPES::ElectionConfigType;
    fn create_election(
        entries: Vec<PeerConfig<TYPES::SignatureKey>>,
        config: TYPES::ElectionConfigType
    ) -> Self;
    fn get_committee_qc_stake_table(
        &self
    ) -> Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>;
    fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey;
    fn get_staked_committee(
        &self,
        view_number: TYPES::Time
    ) -> BTreeSet<TYPES::SignatureKey>;
    fn get_non_staked_committee(
        &self,
        view_number: TYPES::Time
    ) -> BTreeSet<TYPES::SignatureKey>;
    fn get_whole_committee(
        &self,
        view_number: TYPES::Time
    ) -> BTreeSet<TYPES::SignatureKey>;
    fn has_stake(&self, pub_key: &TYPES::SignatureKey) -> bool;
    fn get_stake(
        &self,
        pub_key: &TYPES::SignatureKey
    ) -> Option<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>;
    fn total_nodes(&self) -> usize;
    fn success_threshold(&self) -> NonZeroU64;
    fn failure_threshold(&self) -> NonZeroU64;
    fn upgrade_threshold(&self) -> NonZeroU64;

}

A protocol for determining membership in and participating in a committee.

Required Methods

fn default_election_config( num_nodes_with_stake: u64, num_nodes_without_stake: u64 ) -> TYPES::ElectionConfigType

generate a default election configuration

fn create_election( entries: Vec<PeerConfig<TYPES::SignatureKey>>, config: TYPES::ElectionConfigType ) -> Self

create an election TODO may want to move this to a testableelection trait

fn get_committee_qc_stake_table( &self ) -> Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>

Clone the public key and corresponding stake table for current elected committee

fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey

The leader of the committee for view view_number.

fn get_staked_committee( &self, view_number: TYPES::Time ) -> BTreeSet<TYPES::SignatureKey>

The staked members of the committee for view view_number.

fn get_non_staked_committee( &self, view_number: TYPES::Time ) -> BTreeSet<TYPES::SignatureKey>

The non-staked members of the committee for view view_number.

fn get_whole_committee( &self, view_number: TYPES::Time ) -> BTreeSet<TYPES::SignatureKey>

Get whole (staked + non-staked) committee for view view_number.

fn has_stake(&self, pub_key: &TYPES::SignatureKey) -> bool

Check if a key has stake

fn get_stake( &self, pub_key: &TYPES::SignatureKey ) -> Option<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>

Get the stake table entry for a public key, returns None if the key is not in the table

fn total_nodes(&self) -> usize

Returns the number of total nodes in the committee

fn success_threshold(&self) -> NonZeroU64

Returns the threshold for a specific Membership implementation

fn failure_threshold(&self) -> NonZeroU64

Returns the threshold for a specific Membership implementation

fn upgrade_threshold(&self) -> NonZeroU64

Returns the threshold required to upgrade the network protocol

Object Safety

This trait is not object safe.