Trait hotshot_types::traits::election::Membership

pub trait Membership<TYPES: NodeType>: Clone + Eq + PartialEq + Send + Sync + 'static {
    type StakeTable: Send + Sync;

    fn default_election_config(num_nodes: u64) -> TYPES::ElectionConfigType;
    fn create_election(
        keys: Vec<TYPES::SignatureKey>,
        config: TYPES::ElectionConfigType
    ) -> Self;
    fn get_stake_table(
        &self,
        view_number: TYPES::Time,
        state: &TYPES::StateType
    ) -> Self::StakeTable;
    fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey;
    fn get_committee(
        &self,
        view_number: TYPES::Time
    ) -> BTreeSet<TYPES::SignatureKey>;
    fn make_vote_token(
        &self,
        view_number: TYPES::Time,
        priv_key: &<TYPES::SignatureKey as SignatureKey>::PrivateKey
    ) -> Result<Option<TYPES::VoteTokenType>, ElectionError>;
    fn validate_vote_token(
        &self,
        view_number: TYPES::Time,
        pub_key: TYPES::SignatureKey,
        token: Checked<TYPES::VoteTokenType>
    ) -> Result<Checked<TYPES::VoteTokenType>, ElectionError>;
    fn threshold(&self) -> NonZeroU64;
}

Required Associated Types

type StakeTable: Send + Sync

Required Methods

fn default_election_config(num_nodes: u64) -> TYPES::ElectionConfigType

generate a default election configuration

fn create_election(
keys: Vec<TYPES::SignatureKey>,
config: TYPES::ElectionConfigType
) -> Self

create an election TODO may want to move this to a testableelection trait

fn get_stake_table(
&self,
view_number: TYPES::Time,
state: &TYPES::StateType
) -> Self::StakeTable

Returns the table from the current committed state

fn get_leader(&self, view_number: TYPES::Time) -> TYPES::SignatureKey

fn get_committee(
&self,
view_number: TYPES::Time
) -> BTreeSet<TYPES::SignatureKey>

fn make_vote_token(
&self,
view_number: TYPES::Time,
priv_key: &<TYPES::SignatureKey as SignatureKey>::PrivateKey
) -> Result<Option<TYPES::VoteTokenType>, ElectionError>

Attempts to generate a vote token for self

Returns None if the number of seats would be zero

Errors

TODO tbd

fn validate_vote_token(
&self,
view_number: TYPES::Time,
pub_key: TYPES::SignatureKey,
token: Checked<TYPES::VoteTokenType>
) -> Result<Checked<TYPES::VoteTokenType>, ElectionError>

Checks the claims of a received vote token

Errors

TODO tbd

fn threshold(&self) -> NonZeroU64

Returns the threshold for a specific Membership implementation