Trait hotshot_types::traits::election::QuorumExchangeType

pub trait QuorumExchangeType<TYPES: NodeType, LEAF: LeafType<NodeType = TYPES>, M: NetworkMsg>: ConsensusExchange<TYPES, LEAF, M> {
    fn create_yes_message<I: NodeImplementation<TYPES, Leaf = LEAF>>(
        &self,
        justify_qc_commitment: Commitment<Self::Certificate>,
        leaf_commitment: Commitment<LEAF>,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> ConsensusMessage<TYPES, I>
    where
        <Self as ConsensusExchange<TYPES, LEAF, M>>::Certificate: Committable,
        I::QuorumExchange: ConsensusExchange<TYPES, I::Leaf, Message<TYPES, I>, Vote = QuorumVote<TYPES, LEAF>>;
    fn sign_validating_or_commitment_proposal<I: NodeImplementation<TYPES>>(
        &self,
        leaf_commitment: &Commitment<LEAF>
    ) -> EncodedSignature;
    fn sign_yes_vote(
        &self,
        leaf_commitment: Commitment<LEAF>
    ) -> (EncodedPublicKey, EncodedSignature);
    fn sign_no_vote(
        &self,
        leaf_commitment: Commitment<LEAF>
    ) -> (EncodedPublicKey, EncodedSignature);
    fn sign_timeout_vote(
        &self,
        view_number: TYPES::Time
    ) -> (EncodedPublicKey, EncodedSignature);
    fn create_no_message<I: NodeImplementation<TYPES>>(
        &self,
        justify_qc_commitment: Commitment<QuorumCertificate<TYPES, LEAF>>,
        leaf_commitment: Commitment<LEAF>,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> ConsensusMessage<TYPES, I>
    where
        I::QuorumExchange: ConsensusExchange<TYPES, I::Leaf, Message<TYPES, I>, Vote = QuorumVote<TYPES, LEAF>>;
    fn create_timeout_message<I: NodeImplementation<TYPES>>(
        &self,
        justify_qc: QuorumCertificate<TYPES, LEAF>,
        current_view: TYPES::Time,
        vote_token: TYPES::VoteTokenType
    ) -> ConsensusMessage<TYPES, I>
    where
        I::QuorumExchange: ConsensusExchange<TYPES, I::Leaf, Message<TYPES, I>, Vote = QuorumVote<TYPES, LEAF>>;
}

A ConsensusExchange where participants vote to append items to a log.

Required Methods

fn create_yes_message<I: NodeImplementation<TYPES, Leaf = LEAF>>(
&self,
justify_qc_commitment: Commitment<Self::Certificate>,
leaf_commitment: Commitment<LEAF>,
current_view: TYPES::Time,
vote_token: TYPES::VoteTokenType
) -> ConsensusMessage<TYPES, I>where
<Self as ConsensusExchange<TYPES, LEAF, M>>::Certificate: Committable,
I::QuorumExchange: ConsensusExchange<TYPES, I::Leaf, Message<TYPES, I>, Vote = QuorumVote<TYPES, LEAF>>,

Create a message with a positive vote on validating or commitment proposal.

fn sign_validating_or_commitment_proposal<I: NodeImplementation<TYPES>>(
&self,
leaf_commitment: &Commitment<LEAF>
) -> EncodedSignature

Sign a validating or commitment proposal.

fn sign_yes_vote(
&self,
leaf_commitment: Commitment<LEAF>
) -> (EncodedPublicKey, EncodedSignature)

Sign a positive vote on validating or commitment proposal.

The leaf commitment and the type of the vote (yes) are signed, which is the minimum amount of information necessary for any user of the subsequently constructed QC to check that this node voted Yes on that leaf. The leaf is expected to be reconstructed based on other information in the yes vote.

fn sign_no_vote(
&self,
leaf_commitment: Commitment<LEAF>
) -> (EncodedPublicKey, EncodedSignature)

Sign a neagtive vote on validating or commitment proposal.

The leaf commitment and the type of the vote (no) are signed, which is the minimum amount of information necessary for any user of the subsequently constructed QC to check that this node voted No on that leaf.

fn sign_timeout_vote(
&self,
view_number: TYPES::Time
) -> (EncodedPublicKey, EncodedSignature)

Sign a timeout vote.

We only sign the view number, which is the minimum amount of information necessary for checking that this node timed out on that view.

This also allows for the high QC included with the vote to be spoofed in a MITM scenario, but it is outside our threat model.

fn create_no_message<I: NodeImplementation<TYPES>>(
&self,
justify_qc_commitment: Commitment<QuorumCertificate<TYPES, LEAF>>,
leaf_commitment: Commitment<LEAF>,
current_view: TYPES::Time,
vote_token: TYPES::VoteTokenType
) -> ConsensusMessage<TYPES, I>where
I::QuorumExchange: ConsensusExchange<TYPES, I::Leaf, Message<TYPES, I>, Vote = QuorumVote<TYPES, LEAF>>,

Create a message with a negative vote on validating or commitment proposal.

fn create_timeout_message<I: NodeImplementation<TYPES>>(
&self,
justify_qc: QuorumCertificate<TYPES, LEAF>,
current_view: TYPES::Time,
vote_token: TYPES::VoteTokenType
) -> ConsensusMessage<TYPES, I>where
I::QuorumExchange: ConsensusExchange<TYPES, I::Leaf, Message<TYPES, I>, Vote = QuorumVote<TYPES, LEAF>>,

Create a message with a timeout vote on validating or commitment proposal.