Trait hotshot_types::traits::election::SignedCertificate

pub trait SignedCertificate<SIGNATURE: SignatureKey, TIME, TOKEN, COMMITTABLE>where
    Self: Send + Sync + Clone + Serialize + for<'a> Deserialize<'a>,
    COMMITTABLE: Committable + Serialize + Clone,
    TOKEN: VoteToken,{
    // Required methods
    fn from_signatures_and_commitment(
        view_number: TIME,
        signatures: YesNoSignature<COMMITTABLE, TOKEN>,
        commit: Commitment<COMMITTABLE>
    ) -> Self;
    fn view_number(&self) -> TIME;
    fn signatures(&self) -> YesNoSignature<COMMITTABLE, TOKEN>;
    fn leaf_commitment(&self) -> Commitment<COMMITTABLE>;
    fn set_leaf_commitment(&mut self, commitment: Commitment<COMMITTABLE>);
    fn is_genesis(&self) -> bool;
    fn genesis() -> Self;
}

A certificate of some property which has been signed by a quroum of nodes.

Required Methods

fn from_signatures_and_commitment( view_number: TIME, signatures: YesNoSignature<COMMITTABLE, TOKEN>, commit: Commitment<COMMITTABLE> ) -> Self

Build a QC from the threshold signature and commitment

fn view_number(&self) -> TIME

Get the view number.

fn signatures(&self) -> YesNoSignature<COMMITTABLE, TOKEN>

Get signatures.

fn leaf_commitment(&self) -> Commitment<COMMITTABLE>

Get the leaf commitment.

fn set_leaf_commitment(&mut self, commitment: Commitment<COMMITTABLE>)

Set the leaf commitment.

fn is_genesis(&self) -> bool

Get whether the certificate is for the genesis block.

fn genesis() -> Self

To be used only for generating the genesis quorum certificate; will fail if used anywhere else