Trait hotshot_types::traits::election::SignedCertificate

pub trait SignedCertificate<SIGNATURE: SignatureKey, TIME, TOKEN, LEAF>where
    Self: Send + Sync + Clone + Serialize + for<'a> Deserialize<'a>,
    LEAF: Committable,{
    fn from_signatures_and_commitment(
        view_number: TIME,
        signatures: BTreeMap<EncodedPublicKey, (EncodedSignature, TOKEN)>,
        commit: Commitment<LEAF>
    ) -> Self;
    fn view_number(&self) -> TIME;
    fn signatures(
        &self
    ) -> BTreeMap<EncodedPublicKey, (EncodedSignature, TOKEN)>;
    fn leaf_commitment(&self) -> Commitment<LEAF>;
    fn set_leaf_commitment(&mut self, commitment: Commitment<LEAF>);
    fn is_genesis(&self) -> bool;
    fn genesis() -> Self;
}

Required Methods

fn from_signatures_and_commitment(
view_number: TIME,
signatures: BTreeMap<EncodedPublicKey, (EncodedSignature, TOKEN)>,
commit: Commitment<LEAF>
) -> Self

Build a QC from the threshold signature and commitment

fn view_number(&self) -> TIME

Get the view number.

fn signatures(&self) -> BTreeMap<EncodedPublicKey, (EncodedSignature, TOKEN)>

Get signatures.

fn leaf_commitment(&self) -> Commitment<LEAF>

fn set_leaf_commitment(&mut self, commitment: Commitment<LEAF>)

fn is_genesis(&self) -> bool

Get whether the certificate is for the genesis block.

fn genesis() -> Self

To be used only for generating the genesis quorum certificate; will fail if used anywhere else