Trait hotshot_types::traits::metrics::Metrics

pub trait Metrics: Send + Sync {
    fn create_counter(
        &self,
        label: String,
        unit_label: Option<String>
    ) -> Box<dyn Counter>;
    fn create_gauge(
        &self,
        label: String,
        unit_label: Option<String>
    ) -> Box<dyn Gauge>;
    fn create_histogram(
        &self,
        label: String,
        unit_label: Option<String>
    ) -> Box<dyn Histogram>;
    fn create_label(&self, label: String) -> Box<dyn Label>;
    fn subgroup(&self, subgroup_name: String) -> Box<dyn Metrics>;
}

The metrics type.

Required Methods

fn create_counter(
&self,
label: String,
unit_label: Option<String>
) -> Box<dyn Counter>

Create a Counter with an optional unit_label.

The unit_label can be used to indicate what the unit of the value is, e.g. “kb” or “seconds”

fn create_gauge(
&self,
label: String,
unit_label: Option<String>
) -> Box<dyn Gauge>

Create a Gauge with an optional unit_label.

The unit_label can be used to indicate what the unit of the value is, e.g. “kb” or “seconds”

fn create_histogram(
&self,
label: String,
unit_label: Option<String>
) -> Box<dyn Histogram>

Create a Histogram with an optional unit_label.

The unit_label can be used to indicate what the unit of the value is, e.g. “kb” or “seconds”

fn create_label(&self, label: String) -> Box<dyn Label>

Create a Label.

fn subgroup(&self, subgroup_name: String) -> Box<dyn Metrics>

Create a subgroup with a specified prefix.