hotshot_types::traits
Module network

Network access compatibility

Contains types and traits used by HotShot to abstract over network access

Structs

AsynchronousNetwork

An asynchronous network. Packets may be dropped entirely or delayed for arbitrarily long periods probability that packet is kept = keep_numerator / keep_denominator packet delay is obtained by sampling from a uniform distribution between delay_low_ms and delay_high_ms, inclusive

ChaosNetwork

A chaotic network using all the networking calls

DataRequest

A request for some data that the consensus layer is asking for.

PartiallySynchronousNetwork

An partially synchronous network. Behaves asynchronously until some arbitrary time bound, GST, then synchronously after GST

PerfectNetwork

ideal network

SynchronousNetwork

A synchronous network. Packets may be delayed, but are guaranteed to arrive within timeout ns

Enums

BroadcastDelay

When a message should be broadcast to the network.

NetworkChange

Changes that can occur in the network

NetworkError

Errors that can occur in the network

PushCdnNetworkError

Centralized server specific errors

RequestKind

Underlying data request

ResponseMessage

A response for a request. SequencingMessage is the same as other network messages The kind of message M is determined by what we requested TODO: Put Found content in a Box to make enum smaller

Topic

Used when broadcasting messages

TransmitType

the type of transmission

Traits

ConnectedNetwork

represents a networking implmentration exposes low level API for interacting with a network intended to be implemented for libp2p, the centralized server, and memory network

Id

Trait that bundles what we need from a request ID

NetworkReliability

interface describing how reliable the network is

TestableNetworkingImplementation

Describes additional functionality needed by the test network implementation

ViewMessage

a message

Type Aliases

AsyncGenerator

A channel generator for types that need asynchronous execution