Module hotshot_types::traits::network

Network access compatibility

Contains types and traits used by HotShot to abstract over network access

Structs

AsynchronousNetwork

An asynchronous network. Packets may be dropped entirely or delayed for arbitrarily long periods probability that packet is kept = keep_numerator / keep_denominator packet delay is obtained by sampling from a uniform distribution between delay_low_ms and delay_high_ms, inclusive

CentralizedServerSnafu

SNAFU context selector for the NetworkError::CentralizedServer variant

ChannelSendSnafu

SNAFU context selector for the NetworkError::ChannelSend variant

ChaosNetwork

A chaotic network using all the networking calls

ClientDisconnectedSnafu

SNAFU context selector for the WebServerNetworkError::ClientDisconnected variant

ClientSnafu

SNAFU context selector for the WebServerNetworkError::ClientError variant

CouldNotDeliverSnafu

SNAFU context selector for the NetworkError::CouldNotDeliver variant

DataRequest

A request for some data that the consensus layer is asking for.

EndpointSnafu

SNAFU context selector for the WebServerNetworkError::EndpointError variant

FailedToDeserializeSnafu

SNAFU context selector for the NetworkError::FailedToDeserialize variant

FailedToReceiveSnafu

SNAFU context selector for the PushCdnNetworkError::FailedToReceive variant

FailedToSendSnafu

SNAFU context selector for the PushCdnNetworkError::FailedToSend variant

FailedToSerializeSnafu

SNAFU context selector for the NetworkError::FailedToSerialize variant

IncorrectConsensusDataSnafu

SNAFU context selector for the WebServerNetworkError::IncorrectConsensusData variant

Libp2pMultiSnafu

SNAFU context selector for the NetworkError::Libp2pMulti variant

Libp2pSnafu

SNAFU context selector for the NetworkError::Libp2p variant

MemoryNetworkSnafu

SNAFU context selector for the NetworkError::MemoryNetwork variant

MultipleErrorsSnafu

SNAFU context selector for the NetworkError::MultipleErrors variant

NoBootstrapNodesSpecifiedSnafu

SNAFU context selector for the NetworkError::NoBootstrapNodesSpecified variant

NoMessagesInQueueSnafu

SNAFU context selector for the CentralizedServerNetworkError::NoMessagesInQueue variant

NoSuchNodeSnafu

SNAFU context selector for the NetworkError::NoSuchNode variant

NotFoundSnafu

SNAFU context selector for the NetworkError::NotFound variant

PartiallySynchronousNetwork

An partially synchronous network. Behaves asynchronously until some arbitrary time bound, GST, then synchronously after GST

PerfectNetwork

ideal network

PushCdnNetworkSnafu

SNAFU context selector for the NetworkError::PushCdnNetwork variant

ResponseChannel

Wraps a oneshot channel for responding to requests

ShutDownSnafu

SNAFU context selector for the NetworkError::ShutDown variant

StubSnafu

SNAFU context selector for the MemoryNetworkError::Stub variant

SynchronousNetwork

A synchronous network. Packets may be delayed, but are guaranteed to arrive within timeout ns

TimeoutSnafu

SNAFU context selector for the NetworkError::Timeout variant

UnableToCancelSnafu

SNAFU context selector for the NetworkError::UnableToCancel variant

UnimplementedFeatureSnafu

SNAFU context selector for the NetworkError::UnimplementedFeature variant

WebServerSnafu

SNAFU context selector for the NetworkError::WebServer variant

Enums

CentralizedServerNetworkError

Centralized server specific errors

ConsensusIntentEvent

Events to poll or cancel consensus processes.

MemoryNetworkError

for any errors we decide to add to memory network

NetworkChange

Changes that can occur in the network

NetworkError

Error type for networking

PushCdnNetworkError

Centralized server specific errors

RequestKind

Underlying data request

ResponseMessage

A response for a request. SequencingMessage is the same as other network messages The kind of message M is is determined by what we requested TODO: Put Found content in a Box to make enum smaller

TransmitType

the type of transmission

WebServerNetworkError

Web server specific errors

Traits

ConnectedNetwork

represents a networking implmentration exposes low level API for interacting with a network intended to be implemented for libp2p, the centralized server, and memory network

Id

Trait that bundles what we need from a request ID

NetworkMsg

common traits we would like our network messages to implement

NetworkReliability

interface describing how reliable the network is

TestableNetworkingImplementation

Describes additional functionality needed by the test network implementation

ViewMessage

a message

Type Aliases

AsyncGenerator

A channel generator for types that need asynchronous execution