Module hotshot_types::traits::network

Network access compatibility

Contains types and traits used by HotShot to abstract over network access

Structs

CentralizedServerSnafu

SNAFU context selector for the NetworkError::CentralizedServer variant

ChannelSendSnafu

SNAFU context selector for the NetworkError::ChannelSend variant

ClientDisconnectedSnafu

SNAFU context selector for the WebServerNetworkError::ClientDisconnected variant

ClientSnafu

SNAFU context selector for the WebServerNetworkError::ClientError variant

CouldNotDeliverSnafu

SNAFU context selector for the NetworkError::CouldNotDeliver variant

EndpointSnafu

SNAFU context selector for the WebServerNetworkError::EndpointError variant

FailedToDeserializeSnafu

SNAFU context selector for the NetworkError::FailedToDeserialize variant

FailedToSerializeSnafu

SNAFU context selector for the NetworkError::FailedToSerialize variant

IncorrectConsensusDataSnafu

SNAFU context selector for the WebServerNetworkError::IncorrectConsensusData variant

Libp2pSnafu

SNAFU context selector for the NetworkError::Libp2p variant

MemoryNetworkSnafu

SNAFU context selector for the NetworkError::MemoryNetwork variant

NoMessagesInQueueSnafu

SNAFU context selector for the CentralizedServerNetworkError::NoMessagesInQueue variant

NoSuchNodeSnafu

SNAFU context selector for the NetworkError::NoSuchNode variant

ShutDownSnafu

SNAFU context selector for the NetworkError::ShutDown variant

StubSnafu

SNAFU context selector for the MemoryNetworkError::Stub variant

TimeoutSnafu

SNAFU context selector for the NetworkError::Timeout variant

UnableToCancelSnafu

SNAFU context selector for the NetworkError::UnableToCancel variant

UnimplementedFeatureSnafu

SNAFU context selector for the NetworkError::UnimplementedFeature variant

WebServerSnafu

SNAFU context selector for the NetworkError::WebServer variant

Enums

CentralizedServerNetworkError

Centralized server specific errors

ConsensusIntentEvent

Events to poll or cancel consensus processes.

MemoryNetworkError

for any errors we decide to add to memory network

NetworkChange

Changes that can occur in the network

NetworkError

Error type for networking

TransmitType

the type of transmission

WebServerNetworkError

Web server specific errors

Traits

CommunicationChannel

API for interacting directly with a consensus committee intended to be implemented for both DA and for validating consensus committees

ConnectedNetwork

represents a networking implmentration exposes low level API for interacting with a network intended to be implemented for libp2p, the centralized server, and memory network

NetworkMsg

common traits we would like our network messages to implement

NetworkReliability

interface describing how reliable the network is

TestableChannelImplementation

Describes additional functionality needed by the test communication channel

TestableNetworkingImplementation

Describes additional functionality needed by the test network implementation

ViewMessage

a message