Trait hotshot_types::traits::network::TestableNetworkingImplementation

pub trait TestableNetworkingImplementation<TYPES: NodeType>
where
    Self: Sized,
{
    // Required methods
    fn generator(
        expected_node_count: usize,
        num_bootstrap: usize,
        network_id: usize,
        da_committee_size: usize,
        reliability_config: Option<Box<dyn NetworkReliability>>,
        secondary_network_delay: Duration,
    ) -> AsyncGenerator<Arc<Self>>;
    fn in_flight_message_count(&self) -> Option<usize>;
}

Describes additional functionality needed by the test network implementation

Required Methods

fn generator( expected_node_count: usize, num_bootstrap: usize, network_id: usize, da_committee_size: usize, reliability_config: Option<Box<dyn NetworkReliability>>, secondary_network_delay: Duration, ) -> AsyncGenerator<Arc<Self>>

generates a network given an expected node count

fn in_flight_message_count(&self) -> Option<usize>

Get the number of messages in-flight.

Some implementations will not be able to tell how many messages there are in-flight. These implementations should return None.

Object Safety

This trait is not object safe.