hotshot_types::traits
Module node_implementation

Composite trait for node behavior

This module defines the NodeImplementation trait, which is a composite trait used for describing the overall behavior of a node, as a composition of implementations of the node trait.

Traits

ConsensusTime

Trait for time compatibility needed for reward collection

HasUrls

This trait guarantees that a particular type has urls that can be extracted from it. This trait essentially ensures that the results returned by the AuctionResultsProvider trait includes a list of urls for the builders that HotShot must request from.

NodeImplementation

Node implementation aggregate trait

NodeType

Trait with all the type definitions that are used in the current hotshot setup.

TestableNodeImplementation

extra functions required on a node implementation to be usable by hotshot-testing

Versions

Version information for HotShot