Trait hotshot_types::traits::node_implementation::NodeImplementation

pub trait NodeImplementation<TYPES: NodeType>: Send + Sync + Debug + Clone + 'static + Serialize + for<'de> Deserialize<'de> {
    type Leaf: LeafType<NodeType = TYPES>;
    type Storage: Storage<TYPES, Self::Leaf> + Clone;
    type ConsensusMessage: ConsensusMessageType<TYPES, Self> + Clone + Debug + Send + Sync + 'static + for<'a> Deserialize<'a> + Serialize;
    type Exchanges: ExchangesType<TYPES::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>;
}

Node implementation aggregate trait

This trait exists to collect multiple behavior implementations into one type, to allow HotShot to avoid annoying numbers of type arguments and type patching.

It is recommended you implement this trait on a zero sized type, as HotShotdoes not actually store or keep a reference to any value implementing this trait.

Required Associated Types

type Leaf: LeafType<NodeType = TYPES>

Leaf type for this consensus implementation

type Storage: Storage<TYPES, Self::Leaf> + Clone

Storage type for this consensus implementation

type ConsensusMessage: ConsensusMessageType<TYPES, Self> + Clone + Debug + Send + Sync + 'static + for<'a> Deserialize<'a> + Serialize

Consensus message type.

type Exchanges: ExchangesType<TYPES::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>

Consensus type selected exchanges.

Implements either ValidatingExchangesType or SequencingExchangesType.