Trait hotshot_types::traits::node_implementation::NodeImplementation

pub trait NodeImplementation<TYPES: NodeType>: Send + Sync + Debug + Clone + Eq + Hash + 'static + Serialize + for<'de> Deserialize<'de> {
    type Leaf: LeafType<NodeType = TYPES>;
    type Storage: Storage<TYPES, Self::Leaf> + Clone;
    type ConsensusMessage: ConsensusMessageType<TYPES, Self> + Clone + Debug + Send + Sync + 'static + for<'a> Deserialize<'a> + Serialize;
    type Exchanges: ExchangesType<TYPES::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>;

    // Required method
    fn new_channel_maps(
        start_view: TYPES::Time
    ) -> (ChannelMaps<TYPES, Self>, Option<ChannelMaps<TYPES, Self>>);
}

Node implementation aggregate trait

This trait exists to collect multiple behavior implementations into one type, to allow HotShot to avoid annoying numbers of type arguments and type patching.

It is recommended you implement this trait on a zero sized type, as HotShotdoes not actually store or keep a reference to any value implementing this trait.

Required Associated Types

type Leaf: LeafType<NodeType = TYPES>

Leaf type for this consensus implementation

type Storage: Storage<TYPES, Self::Leaf> + Clone

Storage type for this consensus implementation

type ConsensusMessage: ConsensusMessageType<TYPES, Self> + Clone + Debug + Send + Sync + 'static + for<'a> Deserialize<'a> + Serialize

Consensus message type.

type Exchanges: ExchangesType<TYPES::ConsensusType, TYPES, Self::Leaf, Message<TYPES, Self>>

Consensus type selected exchanges.

Implements either ValidatingExchangesType or SequencingExchangesType.

Required Methods

fn new_channel_maps( start_view: TYPES::Time ) -> (ChannelMaps<TYPES, Self>, Option<ChannelMaps<TYPES, Self>>)

Create channels for sending/recv-ing proposals and votes for quorum and committee exchanges, the latter of which is only applicable for sequencing consensus.