hotshot_types::traits::node_implementation
Trait NodeType

pub trait NodeType:
    Clone
    + Copy
    + Debug
    + Hash
    + PartialEq
    + Eq
    + PartialOrd
    + Ord
    + Default
    + Serialize
    + for<'de> Deserialize<'de>
    + Send
    + Sync
    + 'static {
    type View: ConsensusTime + Display;
    type Epoch: ConsensusTime + Display;
    type AuctionResult: Debug + HasUrls + DeserializeOwned + Default + PartialEq + Eq + Clone + Send + Sync;
    type BlockHeader: BlockHeader<Self>;
    type BlockPayload: BlockPayload<Self, Instance = Self::InstanceState, Transaction = Self::Transaction, ValidatedState = Self::ValidatedState>;
    type SignatureKey: SignatureKey;
    type Transaction: Transaction;
    type InstanceState: InstanceState;
    type ValidatedState: ValidatedState<Self, Instance = Self::InstanceState, Time = Self::View>;
    type Membership: Membership<Self>;
    type BuilderSignatureKey: BuilderSignatureKey;

    const UPGRADE_CONSTANTS: UpgradeConstants = DEFAULT_UPGRADE_CONSTANTS;
}

Trait with all the type definitions that are used in the current hotshot setup.

Provided Associated Constants

const UPGRADE_CONSTANTS: UpgradeConstants = DEFAULT_UPGRADE_CONSTANTS

Constants used to construct upgrade proposals

Required Associated Types

type View: ConsensusTime + Display

The time type that this hotshot setup is using.

This should be the same Time that ValidatedState::Time is using.

type Epoch: ConsensusTime + Display

Same as above but for epoch.

type AuctionResult: Debug + HasUrls + DeserializeOwned + Default + PartialEq + Eq + Clone + Send + Sync

The AuctionSolverResult is a type that holds the data associated with a particular solver run, for a particular view.

type BlockHeader: BlockHeader<Self>

The block header type that this hotshot setup is using.

type BlockPayload: BlockPayload<Self, Instance = Self::InstanceState, Transaction = Self::Transaction, ValidatedState = Self::ValidatedState>

The block type that this hotshot setup is using.

This should be the same block that ValidatedState::BlockPayload is using.

type SignatureKey: SignatureKey

The signature key that this hotshot setup is using.

type Transaction: Transaction

The transaction type that this hotshot setup is using.

This should be equal to BlockPayload::Transaction

type InstanceState: InstanceState

The instance-level state type that this hotshot setup is using.

type ValidatedState: ValidatedState<Self, Instance = Self::InstanceState, Time = Self::View>

The validated state type that this hotshot setup is using.

type Membership: Membership<Self>

Membership used for this implementation

type BuilderSignatureKey: BuilderSignatureKey

The type builder uses to sign its messages

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.