Trait hotshot_types::traits::node_implementation::TestableNodeImplementation

pub trait TestableNodeImplementation<TYPES: NodeType>: NodeImplementation<TYPES> {
    // Required methods
    fn network_generator(
        expected_node_count: usize,
        num_bootstrap_nodes: usize
    ) -> Box<dyn Fn(u64) -> NetworkType<TYPES, Self> + 'static>;
    fn committee_generator(
    ) -> Box<dyn Fn(Arc<NetworkType<TYPES, Self>>) -> <Self::CommitteeExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static>;
    fn quorum_generator(
    ) -> Box<dyn Fn(Arc<NetworkType<TYPES, Self>>) -> <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static>;
    fn state_create_random_transaction(
        state: Option<&TYPES::StateType>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <TYPES::BlockType as Block>::Transaction;
    fn leaf_create_random_transaction(
        leaf: &Self::Leaf,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <TYPES::BlockType as Block>::Transaction;
    fn block_genesis() -> TYPES::BlockType;
    fn txn_count(block: &TYPES::BlockType) -> u64;
    fn construct_tmp_storage() -> Result<Self::Storage, StorageError>;
    fn get_full_state<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn generate_test_key(
        id: u64
    ) -> <TYPES::SignatureKey as SignatureKey>::PrivateKey;
}

extra functions required on a node implementation to be usable by hotshot-testing

Required Methods

fn network_generator( expected_node_count: usize, num_bootstrap_nodes: usize ) -> Box<dyn Fn(u64) -> NetworkType<TYPES, Self> + 'static>

generates a network given an expected node count

fn committee_generator( ) -> Box<dyn Fn(Arc<NetworkType<TYPES, Self>>) -> <Self::CommitteeExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static>

generates a committee communication channel given the network

fn quorum_generator( ) -> Box<dyn Fn(Arc<NetworkType<TYPES, Self>>) -> <Self::QuorumExchange as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Networking + 'static>

generates a quorum communication channel given the network

fn state_create_random_transaction( state: Option<&TYPES::StateType>, rng: &mut dyn RngCore, padding: u64 ) -> <TYPES::BlockType as Block>::Transaction

Get the number of messages in-flight from quorum exchange.

Some implementations will not be able to tell how many messages there are in-flight. These implementations should return None. Get the number of messages in-flight from committee exchange.

Some implementations will not be able to tell how many messages there are in-flight. These implementations should return None. Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn leaf_create_random_transaction( leaf: &Self::Leaf, rng: &mut dyn RngCore, padding: u64 ) -> <TYPES::BlockType as Block>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn block_genesis() -> TYPES::BlockType

generate a genesis block

fn txn_count(block: &TYPES::BlockType) -> u64

the number of transactions in a block

fn construct_tmp_storage() -> Result<Self::Storage, StorageError>

Create ephemeral storage Will be deleted/lost immediately after storage is dropped

Errors

Errors if it is not possible to construct temporary storage.

fn get_full_state<'life0, 'async_trait>( storage: &'life0 Self::Storage ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Return the full internal state. This is useful for debugging.

fn generate_test_key( id: u64 ) -> <TYPES::SignatureKey as SignatureKey>::PrivateKey

The private key of the node id in a test.