Trait hotshot_types::traits::node_implementation::TestableNodeImplementation

pub trait TestableNodeImplementation<CONSENSUS: ConsensusType, TYPES: NodeType<ConsensusType = CONSENSUS>>: NodeImplementation<TYPES> {
    type CommitteeCommChannel;
    type CommitteeNetwork;
    type ViewSyncCommChannel;
    type ViewSyncNetwork;
    type CommitteeElectionConfig;

    // Required methods
    fn network_generator(
        expected_node_count: usize,
        num_bootstrap: usize,
        da_committee_size: usize,
        is_da: bool
    ) -> Box<dyn Fn(u64) -> QuorumNetwork<TYPES, Self> + 'static>
       where QuorumCommChannel<TYPES, Self>: CommunicationChannel<TYPES, Message<TYPES, Self>, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>;
    fn committee_election_config_generator(
        
    ) -> Box<dyn Fn(u64) -> Self::CommitteeElectionConfig + 'static>;
    fn quorum_comm_channel_generator(
        
    ) -> Box<dyn Fn(Arc<QuorumNetwork<TYPES, Self>>) -> QuorumCommChannel<TYPES, Self> + 'static>
       where QuorumCommChannel<TYPES, Self>: CommunicationChannel<TYPES, Message<TYPES, Self>, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>;
    fn committee_comm_channel_generator(
        
    ) -> Box<dyn Fn(Arc<QuorumNetwork<TYPES, Self>>) -> Self::CommitteeCommChannel + 'static>;
    fn view_sync_comm_channel_generator(
        
    ) -> Box<dyn Fn(Arc<QuorumNetwork<TYPES, Self>>) -> Self::ViewSyncCommChannel + 'static>;
    fn state_create_random_transaction(
        state: Option<&TYPES::StateType>,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <TYPES::BlockType as Block>::Transaction;
    fn leaf_create_random_transaction(
        leaf: &Self::Leaf,
        rng: &mut dyn RngCore,
        padding: u64
    ) -> <TYPES::BlockType as Block>::Transaction;
    fn block_genesis() -> TYPES::BlockType;
    fn txn_count(block: &TYPES::BlockType) -> u64;
    fn construct_tmp_storage() -> Result<Self::Storage, StorageError>;
    fn get_full_state<'life0, 'async_trait>(
        storage: &'life0 Self::Storage
    ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn generate_test_key(
        id: u64
    ) -> <TYPES::SignatureKey as SignatureKey>::PrivateKey;
}

extra functions required on a node implementation to be usable by hotshot-testing

Required Associated Types

type CommitteeCommChannel

Communication channel for the DA committee. Only needed for the sequencing consensus.

type CommitteeNetwork

Connected network for the DA committee. Only needed for the sequencing consensus.

type ViewSyncCommChannel

Communication channel for view sync.

type ViewSyncNetwork

Connected network for view sync.

type CommitteeElectionConfig

Election config for the DA committee

Required Methods

fn network_generator( expected_node_count: usize, num_bootstrap: usize, da_committee_size: usize, is_da: bool ) -> Box<dyn Fn(u64) -> QuorumNetwork<TYPES, Self> + 'static>where QuorumCommChannel<TYPES, Self>: CommunicationChannel<TYPES, Message<TYPES, Self>, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>,

Generate a quorum network given an expected node count.

fn committee_election_config_generator( ) -> Box<dyn Fn(u64) -> Self::CommitteeElectionConfig + 'static>

Generates a committee-specific election

fn quorum_comm_channel_generator( ) -> Box<dyn Fn(Arc<QuorumNetwork<TYPES, Self>>) -> QuorumCommChannel<TYPES, Self> + 'static>where QuorumCommChannel<TYPES, Self>: CommunicationChannel<TYPES, Message<TYPES, Self>, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Proposal, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Vote, <QuorumEx<TYPES, Self> as ConsensusExchange<TYPES, Message<TYPES, Self>>>::Membership>,

Generate a quorum communication channel given the network.

fn committee_comm_channel_generator( ) -> Box<dyn Fn(Arc<QuorumNetwork<TYPES, Self>>) -> Self::CommitteeCommChannel + 'static>

Generate a committee communication channel given the network.

fn view_sync_comm_channel_generator( ) -> Box<dyn Fn(Arc<QuorumNetwork<TYPES, Self>>) -> Self::ViewSyncCommChannel + 'static>

Generate a view sync communication channel given the network.

fn state_create_random_transaction( state: Option<&TYPES::StateType>, rng: &mut dyn RngCore, padding: u64 ) -> <TYPES::BlockType as Block>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn leaf_create_random_transaction( leaf: &Self::Leaf, rng: &mut dyn RngCore, padding: u64 ) -> <TYPES::BlockType as Block>::Transaction

Creates random transaction if possible otherwise panics padding is the bytes of padding to add to the transaction

fn block_genesis() -> TYPES::BlockType

generate a genesis block

fn txn_count(block: &TYPES::BlockType) -> u64

the number of transactions in a block

fn construct_tmp_storage() -> Result<Self::Storage, StorageError>

Create ephemeral storage Will be deleted/lost immediately after storage is dropped

Errors

Errors if it is not possible to construct temporary storage.

fn get_full_state<'life0, 'async_trait>( storage: &'life0 Self::Storage ) -> Pin<Box<dyn Future<Output = StorageState<TYPES, Self::Leaf>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Return the full internal state. This is useful for debugging.

fn generate_test_key( id: u64 ) -> <TYPES::SignatureKey as SignatureKey>::PrivateKey

The private key of the node id in a test.