Trait hotshot_types::traits::signature_key::SignatureKey

pub trait SignatureKey: Send + Sync + Clone + Sized + Debug + Hash + Serialize + for<'a> Deserialize<'a> + PartialEq + Eq + PartialOrd + Ord {
    type PrivateKey: Send + Sync + Sized + Clone;
    type StakeTableEntry: Send + Sync + Sized + Clone + Debug + Hash + Eq + Serialize + for<'a> Deserialize<'a>;
    type QCParams: Send + Sync + Sized + Clone + Debug + Hash;
    type QCType: Send + Sync + Sized + Clone + Debug + Hash + PartialEq + Eq + Serialize + for<'a> Deserialize<'a>;

    // Required methods
    fn validate(&self, signature: &EncodedSignature, data: &[u8]) -> bool;
    fn sign(private_key: &Self::PrivateKey, data: &[u8]) -> EncodedSignature;
    fn from_private(private_key: &Self::PrivateKey) -> Self;
    fn to_bytes(&self) -> EncodedPublicKey;
    fn from_bytes(bytes: &EncodedPublicKey) -> Option<Self>;
    fn generated_from_seed_indexed(
        seed: [u8; 32],
        index: u64
    ) -> (Self, Self::PrivateKey);
    fn get_stake_table_entry(&self, stake: u64) -> Self::StakeTableEntry;
    fn get_public_parameter(
        stake_entries: Vec<Self::StakeTableEntry>,
        threshold: U256
    ) -> Self::QCParams;
    fn check(
        real_qc_pp: &Self::QCParams,
        data: &[u8],
        qc: &Self::QCType
    ) -> bool;
    fn get_sig_proof(
        signature: &Self::QCType
    ) -> (<BLSOverBN254CurveSignatureScheme as SignatureScheme>::Signature, BitVec);
    fn assemble(
        real_qc_pp: &Self::QCParams,
        signers: &BitSlice,
        sigs: &[<BLSOverBN254CurveSignatureScheme as SignatureScheme>::Signature]
    ) -> Self::QCType;
}

Trait for abstracting public key signatures

Required Associated Types

type PrivateKey: Send + Sync + Sized + Clone

The private key type for this signature algorithm

type StakeTableEntry: Send + Sync + Sized + Clone + Debug + Hash + Eq + Serialize + for<'a> Deserialize<'a>

The type of the entry that contain both public key and stake value

type QCParams: Send + Sync + Sized + Clone + Debug + Hash

The type of the quorum certificate parameters used for assembled signature

type QCType: Send + Sync + Sized + Clone + Debug + Hash + PartialEq + Eq + Serialize + for<'a> Deserialize<'a>

The type of the assembled qc: assembled signature + BitVec

Required Methods

fn validate(&self, signature: &EncodedSignature, data: &[u8]) -> bool

Validate a signature

fn sign(private_key: &Self::PrivateKey, data: &[u8]) -> EncodedSignature

Produce a signature

fn from_private(private_key: &Self::PrivateKey) -> Self

Produce a public key from a private key

fn to_bytes(&self) -> EncodedPublicKey

Serialize a public key to bytes

fn from_bytes(bytes: &EncodedPublicKey) -> Option<Self>

Deserialize a public key from bytes

fn generated_from_seed_indexed( seed: [u8; 32], index: u64 ) -> (Self, Self::PrivateKey)

Generate a new key pair

fn get_stake_table_entry(&self, stake: u64) -> Self::StakeTableEntry

get the stake table entry from the public key and stake value

fn get_public_parameter( stake_entries: Vec<Self::StakeTableEntry>, threshold: U256 ) -> Self::QCParams

get the public parameter for the assembled signature checking

fn check(real_qc_pp: &Self::QCParams, data: &[u8], qc: &Self::QCType) -> bool

check the quorum certificate for the assembled signature

fn get_sig_proof( signature: &Self::QCType ) -> (<BLSOverBN254CurveSignatureScheme as SignatureScheme>::Signature, BitVec)

get the assembled signature and the BitVec separately from the assembled signature

fn assemble( real_qc_pp: &Self::QCParams, signers: &BitSlice, sigs: &[<BLSOverBN254CurveSignatureScheme as SignatureScheme>::Signature] ) -> Self::QCType

assemble the signature from the partial signature and the indication of signers in BitVec