Module hotshot_types::utils

Utility functions, type aliases, helper structs and enum definitions.

Structs

BuilderCommitment

Commitment that builders use to sign block options. A thin wrapper around a Sha256 digest.

RoundFinishedEvent

A struct containing information about a finished round.

View

This exists so we can perform state transitions mutably

Enums

EpochTransitionIndicator

A helper enum to indicate whether a node is in the epoch transition A node is in epoch transition when its high QC is for the last block in an epoch

Terminator

Whether or not to stop inclusively or exclusively when walking

ViewInner

A view’s state

Functions

bincode_opts

For the wire format, we use bincode with the following options:

epoch_from_block_number

Returns an epoch number given a block number and an epoch height

genesis_epoch_from_version

Returns Some(0) if epochs are enabled by V::Base, otherwise returns None

is_epoch_root

Returns true if the given block number is the third from the last in the epoch based on the given epoch height.

is_last_block_in_epoch

Returns true if the given block number is the last in the epoch based on the given epoch height.

mnemonic

A function for generating a cute little user mnemonic from a hash

option_epoch_from_block_number

Returns an Option based on a boolean condition of whether or not epochs are enabled, a block number, and the epoch height. If epochs are disabled or the epoch height is zero, returns None.

root_block_in_epoch

Returns the block number of the epoch root in the given epoch

Type Aliases

LeafCommitment

The hash of a leaf.

Sha256Digest 🔒

Type alias for byte array of SHA256 digest length

StateAndDelta

Optional validated state and state delta.