hotshot_types
Module vote

Vote, Accumulator, and Certificate Types

Structs

VoteAccumulator

Accumulates votes until a certificate is formed. This implementation works for all simple vote and certificate pairs

Traits

Certificate

The certificate formed from the collection of signatures a committee. The committee is defined by the Membership associated type. The votes all must be over the Commitment associated type.

HasViewNumber

Any type that is associated with a view

Vote

A simple vote that has a signer and commitment to the data voted on.

Type Aliases

SignersMap 🔒

Mapping of vote commitment to signatures and bitvec

VoteMap2 🔒

Mapping of commitments to vote tokens by key.