Trait hotshot_types::vote::Certificate

pub trait Certificate<TYPES: NodeType, T>: HasViewNumber<TYPES> {
    type Voteable: Voteable<TYPES>;
    type Threshold: Threshold<TYPES>;

    // Required methods
    fn create_signed_certificate<V: Versions>(
        vote_commitment: Commitment<VersionedVoteData<TYPES, Self::Voteable, V>>,
        data: Self::Voteable,
        sig: <TYPES::SignatureKey as SignatureKey>::QcType,
        view: TYPES::View,
    ) -> Self;
    fn is_valid_cert<V: Versions>(
        &self,
        stake_table: Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>,
        threshold: NonZeroU64,
        upgrade_lock: &UpgradeLock<TYPES, V>,
    ) -> impl Future<Output = Result<()>>;
    fn threshold<MEMBERSHIP: Membership<TYPES>>(
        membership: &MEMBERSHIP,
        epoch: Option<TYPES::Epoch>,
    ) -> u64;
    fn stake_table<MEMBERSHIP: Membership<TYPES>>(
        membership: &MEMBERSHIP,
        epoch: Option<TYPES::Epoch>,
    ) -> Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>;
    fn total_nodes<MEMBERSHIP: Membership<TYPES>>(
        membership: &MEMBERSHIP,
        epoch: Option<TYPES::Epoch>,
    ) -> usize;
    fn stake_table_entry<MEMBERSHIP: Membership<TYPES>>(
        membership: &MEMBERSHIP,
        pub_key: &TYPES::SignatureKey,
        epoch: Option<TYPES::Epoch>,
    ) -> Option<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>;
    fn data(&self) -> &Self::Voteable;
    fn data_commitment<V: Versions>(
        &self,
        upgrade_lock: &UpgradeLock<TYPES, V>,
    ) -> impl Future<Output = Result<Commitment<VersionedVoteData<TYPES, Self::Voteable, V>>>>;
}

The certificate formed from the collection of signatures a committee. The committee is defined by the Membership associated type. The votes all must be over the Commitment associated type.

Required Associated Types

type Voteable: Voteable<TYPES>

The data commitment this certificate certifies.

type Threshold: Threshold<TYPES>

Threshold Functions

Required Methods

fn create_signed_certificate<V: Versions>( vote_commitment: Commitment<VersionedVoteData<TYPES, Self::Voteable, V>>, data: Self::Voteable, sig: <TYPES::SignatureKey as SignatureKey>::QcType, view: TYPES::View, ) -> Self

Build a certificate from the data commitment and the quorum of signers

fn is_valid_cert<V: Versions>( &self, stake_table: Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>, threshold: NonZeroU64, upgrade_lock: &UpgradeLock<TYPES, V>, ) -> impl Future<Output = Result<()>>

Checks if the cert is valid in the given epoch

fn threshold<MEMBERSHIP: Membership<TYPES>>( membership: &MEMBERSHIP, epoch: Option<TYPES::Epoch>, ) -> u64

Returns the amount of stake needed to create this certificate

fn stake_table<MEMBERSHIP: Membership<TYPES>>( membership: &MEMBERSHIP, epoch: Option<TYPES::Epoch>, ) -> Vec<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>

Get Stake Table from Membership implementation.

fn total_nodes<MEMBERSHIP: Membership<TYPES>>( membership: &MEMBERSHIP, epoch: Option<TYPES::Epoch>, ) -> usize

Get Total Nodes from Membership implementation.

fn stake_table_entry<MEMBERSHIP: Membership<TYPES>>( membership: &MEMBERSHIP, pub_key: &TYPES::SignatureKey, epoch: Option<TYPES::Epoch>, ) -> Option<<TYPES::SignatureKey as SignatureKey>::StakeTableEntry>

Get StakeTableEntry from Membership implementation.

fn data(&self) -> &Self::Voteable

Get the commitment which was voted on

fn data_commitment<V: Versions>( &self, upgrade_lock: &UpgradeLock<TYPES, V>, ) -> impl Future<Output = Result<Commitment<VersionedVoteData<TYPES, Self::Voteable, V>>>>

Get the vote commitment which the votes commit to

Object Safety

This trait is not object safe.