Trait hotshot_types::vote::Vote

pub trait Vote<TYPES: NodeType>: HasViewNumber<TYPES> {
    type Commitment: Voteable;

    // Required methods
    fn signature(
        &self,
    ) -> <TYPES::SignatureKey as SignatureKey>::PureAssembledSignatureType;
    fn date(&self) -> &Self::Commitment;
    fn data_commitment(&self) -> Commitment<Self::Commitment>;
    fn signing_key(&self) -> TYPES::SignatureKey;
}

A simple vote that has a signer and commitment to the data voted on.

Required Associated Types

type Commitment: Voteable

Type of data commitment this vote uses.

Required Methods

fn signature( &self, ) -> <TYPES::SignatureKey as SignatureKey>::PureAssembledSignatureType

Get the signature of the vote sender

fn date(&self) -> &Self::Commitment

Gets the data which was voted on by this vote

fn data_commitment(&self) -> Commitment<Self::Commitment>

Gets the Data commitment of the vote

fn signing_key(&self) -> TYPES::SignatureKey

Gets the public signature key of the votes creator/sender